% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddFarmers.R
\name{AddFarmers}
\alias{AddFarmers}
\title{R6 Class for adding farmer information in OFPE database.}
\description{
R6 class that takes a vector of farmer names to add to the
database. Adds a unique ID field for each farmer that is used for
internal relationships. If a farmer is already present in the database,
it will be skipped. This follows the ManageDB interface and includes
a setup and execute method that is called from the ManageDB class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{farmers}}{Vector of farmer names to add to the database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AddFarmers$new()}}
\item \href{#method-setup}{\code{AddFarmers$setup()}}
\item \href{#method-execute}{\code{AddFarmers$execute()}}
\item \href{#method-clone}{\code{AddFarmers$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an AddFarmers object. The database connection and a vector of
farmer names are used to upload farmer information to the database.
Instantiates the class for running the setup and execute methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarmers$new(dbCon, action_item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{action_item}}{List of inputs to the AddFarmers method, from the
list passed into the ManageDB class. This includes a vector of farmer
names to add to the 'all_farms.farmers' table of an OFPE formatted
database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new initialized 'AddFarmers' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Prepares the farmer data prior to uploading to the database.
Queries the database for existing farmer information, and if found
matches existing farmer ID's to farmer names where needed, else
provides a new farmer ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarmers$setup()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Prepared R6 class for uploading farmer information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Executes the upload of prepared farmer information to the database in the
'all_farms' schema and the 'farmers' table. If a farmer is already
present with the same name and ID, it will be skipped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarmers$execute()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Farmer data uploaded into 'all_farms.farmers' table in the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarmers$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
