% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddFarms.R
\name{AddFarms}
\alias{AddFarms}
\title{R6 Class for adding farm information in OFPE database.}
\description{
R6 class that takes a vector of farm names and a path
to the location of farm boundary shapefiles. Uploads boundaries to the
database. Adds a unique ID field for each farm that is used for
internal relationships and identifies the associated farmer from the
names of the farms. If a farm is already present in the database,
it will be replaced with the same farm ID. This follows the ManageDB
interface and includes a setup and execute method that is called from
the ManageDB class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{farm_names}}{Vector of farm names to add to the database.}

\item{\code{farm_farmer_names}}{Vector of farmer names reflecting the
owner of the farm names supplied.}

\item{\code{farm_path}}{File path to the location of farm boundary shapefiles.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AddFarms$new()}}
\item \href{#method-setup}{\code{AddFarms$setup()}}
\item \href{#method-execute}{\code{AddFarms$execute()}}
\item \href{#method-clone}{\code{AddFarms$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an AddFarms object. The database connection and a vector of
farm names and farmer names are used to upload farm information
to the database with the appropriate owner information. The farm_path
argument indicates the file path in which to follow for the farm boundary
shapefiles that are to be uploaded to the database. Instantiates the class
for running the setup and execute methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarms$new(dbCon, farm_names, farm_farmer_names, farm_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{farm_names}}{Vector of farm names to add to the database.}

\item{\code{farm_farmer_names}}{Vector of farmer names reflecting the
owner of the farm names supplied.}

\item{\code{farm_path}}{File path to the location of farm boundary shapefiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new initialized 'AddFarms' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Prepares the farm data prior to uploading to the database. Imports
shapefiles from the farm path and puts them into a list with the
corresponding farmer to be uploaded to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarms$setup()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Prepared R6 class for uploading farmer information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Executes the upload of prepared farm information to the database in the
'all_farms' schema and the 'farms' table. If a farm boundary is already
present with the same farm name and farmer that owns it, the entry in the
database is replaced.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarms$execute()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Farm data uploaded into the 'all_farms.farms' table in the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFarms$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
