% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddFields.R
\name{AddFields}
\alias{AddFields}
\title{R6 Class for adding field information in OFPE database.}
\description{
R6 class that takes a vector of field names and a path
to the location of field boundary shapefiles. Uploads boundaries to the
database. Adds a unique ID field for each field that is used for
internal relationships and identifies the associated farmer with the fields.
If a field is already present in the database,
it will be replaced with the same field ID. This follows the ManageDB
interface and includes a setup and execute method that is called from
the ManageDB class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object, see DBCon class.}

\item{\code{field_names}}{Vector of field names to add to the database.}

\item{\code{field_farmer_names}}{Vector of farmer names reflecting the
owner of the field names supplied.}

\item{\code{field_path}}{File path to the location of field boundary shapefiles.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AddFields$new()}}
\item \href{#method-setup}{\code{AddFields$setup()}}
\item \href{#method-execute}{\code{AddFields$execute()}}
\item \href{#method-clone}{\code{AddFields$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an AddFields object. The database connection and a vector of
field names and farmer names are used to upload field information
to the database with the appropriate owner information. The field_path
argument indicates the file path in which to follow for the field boundary
shapefiles that are to be uploaded to the database. Instantiates the class
for running the setup and execute methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFields$new(dbCon, field_names, field_farmer_names, field_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object, see DBCon class.}

\item{\code{field_names}}{Vector of field names to add to the database.}

\item{\code{field_farmer_names}}{Vector of farmer names reflecting the
owner of the field names supplied.}

\item{\code{field_path}}{File path to the location of field boundary shapefiles.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new initialized 'AddFarms' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
Prepares the field data prior to uploading to the database. Imports
shapefiles from the field path and puts them into a list with the
corresponding farmer to be uploaded to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFields$setup()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Prepared R6 class for uploading farmer information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Executes the upload of prepared field information to the database in the
'all_farms' schema and the 'fields' table. If a field boundary is already
present with the same field name and farmer that owns it, the entry in the
database is replaced.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFields$execute()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Farm data uploaded into the 'all_farms.fields' table in the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddFields$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
