% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggGEE.R
\name{AggGEE}
\alias{AggGEE}
\title{R6 Class for aggregating GEE data}
\description{
R6 Class for executing the OFPE data aggregation process that
consolidates data from Google Earth Engine to on-farm data. This is used
within the 'AggDat' R6 class.
}
\seealso{
\code{\link{DBCon}} for the database connection class,
\code{\link{AggDat}} for the class responsible for aggregating on-farm data,
\code{\link{AggInputs}} for the inputs required for the aggregation
process.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{aggInputs}}{An object of the 'AggInputs' class containaing the options
for aggregating data. This includes the database connection.}

\item{\code{farmidx}}{The ID of the farm in which the field falls within.}

\item{\code{farmeridx}}{The ID of the farmer that owns the farm that the field
falls within.}

\item{\code{PY}}{The year prior to the selected year of interest (cy_resp).
This differs from the previous crop year used prior.}

\item{\code{PY2}}{The year two years prior to the selected year of interest
(cy_resp).}

\item{\code{labels}}{The labels of the GEE covariates that are collected and
aggregated.}

\item{\code{type}}{The type of data corresponding to each respective label.}

\item{\code{SOURCE}}{The SOURCE of data corresponding to each respective label.}

\item{\code{year}}{The year of data to gather, corresponding to each respective
label.}

\item{\code{loy}}{The length of year for which to gather data for each respective
label.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AggGEE$new()}}
\item \href{#method-aggregateGEE}{\code{AggGEE$aggregateGEE()}}
\item \href{#method-.getGEEdata}{\code{AggGEE$.getGEEdata()}}
\item \href{#method-clone}{\code{AggGEE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object for executing the process of aggregating
Google Earth Engine data to OFPE data. Adds necessary columns to
the temporary aggregated table,  sets up the necessary options for
gathering and extracting GEE data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggGEE$new(aggInputs, farmidx, farmeridx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aggInputs}}{An 'AggInputs' R6 class with the user's aggregation
options.}

\item{\code{farmidx}}{The ID of the farm in which the field falls within.}

\item{\code{farmeridx}}{The ID of the farmer that owns the farm that the field
falls within.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An initialized 'AggGEE' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-aggregateGEE"></a>}}
\if{latex}{\out{\hypertarget{method-aggregateGEE}{}}}
\subsection{Method \code{aggregateGEE()}}{
Method for executing the aggregation of GEE data to the on-farm data
within the OFPE database. First, gather the environmental variables
that do not change year to year. Gather from 2015 to make sure they
are there. Gather Daymet V3 and GRIDMET data, when possible, for
both the current and previous year. Get current and previous year
precipitation and growing degree day data. Gather vegetation index
data from the current, previous, and two years prior. Do this for
NDVI, NDRE, and CIRE indices. Additionally, when applicable, gather
Landsat 8 and Sentinel 2 data. Because Landsat does not have the
band frequencies to calculate NDRE and CIRE these are only available
when Sentinel 2 data is available. If available, gather SMAP data
from the current and previous year. Get the surface soil moisture
and subsurface soil moisture from SMAP.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggGEE$aggregateGEE()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments necessary because identified in class
instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
GEE data aggregated to the temporary aggregated table
in the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.getGEEdata"></a>}}
\if{latex}{\out{\hypertarget{method-.getGEEdata}{}}}
\subsection{Method \code{.getGEEdata()}}{
Method for aggregating the Google Earth Engine data to the on-farm
aggregated data. Identifies the appropriate data from the 'all_farms.
gee' schema and extracts the values of each raster to each point in
the field. All labels, type, SOURCE etc. are identified in the GEE
data original filenames exported from GEE. The dot indicates that
this function would be private if not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggGEE$.getGEEdata(
  label,
  type,
  SOURCE,
  year,
  loy,
  db,
  farmidx,
  farmeridx,
  farmername
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{The labels for the GEE data to aggregate.}

\item{\code{type}}{The type of data to aggregate.}

\item{\code{SOURCE}}{The SOURCE of the data. (i.e. Landsat vs. Sentinel etc.).}

\item{\code{year}}{The year of the data to gather.}

\item{\code{loy}}{The length of year to get data from (i.e. 'mar' vs. 'full').}

\item{\code{db}}{Connection to an OFPE database.}

\item{\code{farmidx}}{ID of the farm that the field falls within.}

\item{\code{farmeridx}}{ID of the farmer that owns the farm that the field
falls within.}

\item{\code{farmername}}{Name of the farmer that owns the field for aggregation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data in temporary aggregated table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggGEE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
