% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggInputs.R
\name{AggInputs}
\alias{AggInputs}
\title{R6 Class for storing inputs to the 'AggDat' class}
\description{
R6 class for for storing information needed for the 'AggDat'
class that executes the aggregation methods. These include the field and year
to aggregate data for, the locations to aggregate data to, the response variable
to use, the experimental variable, the length of year to gather for, etc.

Inputs can be supplied directly to this class during instantiation, however
this is NOT recommended except for advanced users. It is recommended that the
user supplies the database connection and uses the interactive selection
methods to select user inputs.

This class is passed to the 'AggDat' class that executes the methods for
aggregating data and storing in the database. Most methods are executed
in the database.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{bboxImport}}{Yes/No, will user be uploading their own field boundary?
Used for spatially querying the database for intersecting data.}

\item{\code{bboxLocation}}{Only relevant if bboxImport == "Yes". Is the location
of the shapefile containing the field boundary to use for spatial queries.}

\item{\code{FIELDNAME}}{Name of the field to aggregate data for. Selected from
the 'all_farms.fields' table of an OFPE formatted database.}

\item{\code{FARMERNAME}}{Name of the farmer that owns the selected field.}

\item{\code{RESPVAR}}{Response variable to aggregate data for, select/input
'Yield', 'Protein', 'Satellite'. 'Satellite' data aggregates only
remotely sensed data and does not include any on-farm collected data.}

\item{\code{EXPVAR}}{Experimental variable to aggregate data for, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{CY_RESP}}{The year of interest of the selected response variable.
This is considered the 'current year' (CY), and differs from the 'previous
year' (PY), which is the latest year before the CY during which the field
was cropped. This is separate from the CY_EXP variable because in some
instances, applied data was categorized as or applied in the year prior
(i.e. WW seeding rates occur in fall of the year before harvest in the
following August).}

\item{\code{PY_RESP}}{The year prior to the selected CY that a crop was harvested
in the specified field. This iss the latest year before the CY during which
the field was cropped. If you do not data from any previous year, you can
provide a year for labeling and annotations sake in output figures.}

\item{\code{CY_EXP}}{The year of interest of the selected experimental variable.
This is the year that the experimental variable was applied to grow the
crop in the selected 'CY_RESP' year. This is separate from the CY_RESP
variable because in some instances, applied data was categorized as or
applied in the year prior (i.e. WW seeding rates occur in fall of the year
before harvest in the following August).}

\item{\code{PY_EXP}}{The application year of the experimental variable that was
used to grow the crop in the 'PY_RESP' year. If you do not data from any
previous year, you can provide a year for labeling and annotations sake in
output figures.}

\item{\code{GRID}}{Determines the location of the aggregated data. Either select
'Grid' or 'Observed'. 'Grid' aggregates data to the centroids of a 10m grid
laid across the field, while 'Observed' aggregates data to the locations
of the observed response variable (yield or protein). Note that when
'Satellite' is selected there are no observed points to use, so the 'Grid'
option is selected by default.}

\item{\code{DAT_USED}}{Option for the length of year to use for CY data. In
winter wheat conventional systems and organic spring wheat systems,
there is a decision point around March 30th at which farmers must make
decisions on their fertilizer or seeding rates, respectively. When making
these decisions, farmers will only have data up to that decision point. This
options determines the covariate data aggregated for the CY, where
'Decision Point' aggregates data for the CY from January 1st to March 30th
of that year, while 'Full Year' aggregates data for the CY from January 1st
to December 31st. This is a residual special use case for Paul Hegedus' 2020
ICPA work.}

\item{\code{CY_RESP_FILES}}{Vector of names. Based on the user's selected field
and year, the database can be queried for the original file names of the
data uploaded to the database. The user selects the correct file name that
corresponds to the response variable data in the specified field and year.
In this case the 'CY_RESP' year. Multiple files are allowed for selection
because of the cases in which multiple files correspond to the data in the
given year and field (i.e. if harvest took multiple days etc.).}

\item{\code{PY_RESP_FILES}}{Vector of names. Based on the user's selected field
and year, the database can be queried for the original file names of the
data uploaded to the database. The user selects the correct file name that
corresponds to the response variable data in the specified field and year.
In this case the 'PY_RESP' year. Multiple files are allowed for selection
because of the cases in which multiple files correspond to the data in the
given year and field (i.e. if harvest took multiple days etc.). If you do
not have data available from a previous year and are not using the
interactive input selectors, simply input "None", otherwise when using the
interactive methods this is handled for you.}

\item{\code{CY_EXP_FILES}}{Data.frame with 2 columns, 'orig_file' and 'table'.
Based on the user's selected field and year, the database can be queried
for the original file names of the data uploaded to the database. The user
selects the correct file name that corresponds to the experimental variable
data in the specified field and year. In this case, the 'CY_EXP' year.
Multiple files are allowed for selection because of the cases in which
multiple files correspond to the data in the given year and field (i.e.
if application took multiple days etc.). The 'orig_file' column contains
the name of the original filename uploaded to the database with the raw
data and 'table' contains the table within the farmer schema that the
data is housed. This is because some as-applied experimental data are
point vectors and some are polygons.}

\item{\code{PY_EXP_FILES}}{Data.frame with 2 columns, 'orig_file' and 'table'.
Based on the user's selected field and year, the database can be queried
for the original file names of the data uploaded to the database. The user
selects the correct file name that corresponds to the experimental variable
data in the specified field and year. In this case, the 'PY_EXP' year.
Multiple files are allowed for selection because of the cases in which
multiple files correspond to the data in the given year and field (i.e. if
application took multiple days etc.). The 'orig_file' column contains
the name of the original filename uploaded to the database with the raw
data and 'table' contains the table within the farmer schema that the
data is housed. This is because some as-applied experimental data are
point vectors and some are polygons. If no data from the desired year
is available, put "None" in the 'orig_file' column.}

\item{\code{saveInDB}}{Yes/No. Logical, whether to save the aggregated data into
the OFPE formatted database. Not an option if you are supplying your own
boundary for which to aggregate data.}

\item{\code{export}}{Yes/No. Logical, whether to export the aggregated data as
a '.csv' file. If yes, the user will need to provide the 'exportName'.}

\item{\code{exportName}}{If exporting the aggregated data as a '.csv', the user
needs to specify the name of the file to export. This includes the file
path.}

\item{\code{CY_RESP_COL}}{Data.frame with 3 columns, 'RESP', 'DIST', and
'orig_file'. This data.frame contains the column containing data for
the selected response variable ('RESP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional (can be NA) and optionally used during a cleaning
step where points are removed if they are more than 4SD from the mean
distance between observations as this indicates that the equipment was
moving at an abnormal speed and potentially resulting in erroneous
measurements. The 'orig_file' is the same as those selected in the
'CY_RESP_FILES' vector.}

\item{\code{PY_RESP_COL}}{Data.frame with 3 columns, 'RESP', 'DIST', and
'orig_file'. This data.frame contains the column containing data for
the selected response variable ('RESP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional and optionally used during a cleaning step where
points are removed if they are more than 4SD from the mean distance
between observations as this indicates that the equipment was moving
at an abnormal speed and potentially resulting in erroneous measurements.
The 'orig_file' is the same as those selected in the 'PY_RESP_FILES' vector.}

\item{\code{CY_EXP_COL}}{Data.frame with 4 columns, 'EXP', 'DIST', 'PRODUCT',
and 'orig_file'. This data.frame contains the column containing data for
the selected experimental variable ('EXP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional (can be NA) and optionally used during a cleaning
step where points are removed if they are more than 4SD from the mean
distance between observations as this indicates that the equipment was
moving at an abnormal speed and potentially resulting in erroneous
measurements. The 'PRODUCT' column is also optional, and not relevant
for seeding rate data, and used if there is a column corresponding to
the product applied. When left blank, it is assumed that the 'EXP' column
contains data in lbs/acre and the user is not given an option to provide
a conversion rate. It is good practice to always select a column (even if
no specific 'product' column) here to explicitly state a conversion factor.
The 'orig_file' is the same as those selected in the 'CY_EXP_FILES' table.}

\item{\code{PY_EXP_COL}}{Data.frame with 4 columns, 'EXP', 'DIST', 'PRODUCT',
and 'orig_file'. This data.frame contains the column containing data for
the selected experimental variable ('EXP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional (can be NA) and optionally used during a cleaning
step where points are removed if they are more than 4SD from the mean
distance between observations as this indicates that the equipment was
moving at an abnormal speed and potentially resulting in erroneous
measurements. The 'PRODUCT' column is also optional, and not relevant
for seeding rate data, and used if there is a column corresponding to
the product applied. When left blank, it is assumed that the 'EXP' column
contains data in lbs/acre and the user is not given an option to provide
a conversion rate. It is good practice to always select a column (even if
no specific 'product' column) here to explicitly state a conversion factor.
The 'orig_file' is the same as those selected in the 'PY_EXP_FILES' table.}

\item{\code{CY_EXP_CONV}}{Data.frame with 3 columns, 'FORMULA', 'CONVERSION',
and 'orig_file'. Based on the users selection of 'PRODUCT' in the
CY_EXP_COL data.frame, the formula from the 'PRODUCT' column is extracted
and used to ask the user the desired conversion factor from the product
applied to lbs per acre. Again, this is only really applicable for fertilzier
rates unless seeding rates are reported in units besides lbs per acre. The
'orig_file' is the same as those selected in the 'CY_EXP_FILES' table.}

\item{\code{PY_EXP_CONV}}{Data.frame with 3 columns, 'FORMULA', 'CONVERSION',
and 'orig_file'. Based on the users selection of 'PRODUCT' in the
CY_EXP_COL data.frame, the formula from the 'PRODUCT' column is extracted
and used to ask the user the desired conversion factor from the product
applied to lbs per acre. Again, this is only really applicable for fertilzier
rates unless seeding rates are reported in units besides lbs per acre. The
'orig_file' is the same as those selected in the 'CY_EXP_FILES' table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AggInputs$new()}}
\item \href{#method-selectInputs}{\code{AggInputs$selectInputs()}}
\item \href{#method-clone}{\code{AggInputs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object for storing aggregation inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggInputs$new(
  dbCon,
  bboxImport = NULL,
  bboxLocation = NULL,
  FIELDNAME = NULL,
  FARMERNAME = NULL,
  RESPVAR = NULL,
  EXPVAR = NULL,
  CY_RESP = NULL,
  PY_RESP = NULL,
  CY_EXP = NULL,
  PY_EXP = NULL,
  GRID = NULL,
  DAT_USED = NULL,
  CY_RESP_FILES = NULL,
  PY_RESP_FILES = NULL,
  CY_EXP_FILES = NULL,
  PY_EXP_FILES = NULL,
  saveInDB = NULL,
  export = NULL,
  exportName = NULL,
  CY_RESP_COL = NULL,
  PY_RESP_COL = NULL,
  CY_EXP_COL = NULL,
  PY_EXP_COL = NULL,
  CY_EXP_CONV = NULL,
  PY_EXP_CONV = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{bboxImport}}{Yes/No, will user be uploading their own field boundary?
Used for spatially querying the database for intersecting data.}

\item{\code{bboxLocation}}{Only relevant if bboxImport == "Yes". Is the location
of the shapefile containing the field boundary to use for spatial queries.}

\item{\code{FIELDNAME}}{Name of the field to aggregate data for. Selected from
the 'all_farms.fields' table of an OFPE formatted database.}

\item{\code{FARMERNAME}}{Name of the farmer that owns the selected field.}

\item{\code{RESPVAR}}{Response variable to aggregate data for, select/input
'Yield', 'Protein', 'Satellite'. 'Satellite' data aggregates only
remotely sensed data and does not include any on-farm collected data.}

\item{\code{EXPVAR}}{Experimental variable to aggregate data for, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{CY_RESP}}{The year of interest of the selected response variable.
This is considered the 'current year' (CY), and differs from the 'previous
year' (PY), which is the latest year before the CY during which the field
was cropped. This is separate from the CY_EXP variable because in some
instances, applied data was categorized as or applied in the year prior
(i.e. WW seeding rates occur in fall of the year before harvest in the
following August).}

\item{\code{PY_RESP}}{The year prior to the selected CY that a crop was harvested
in the specified field. This iss the latest year before the CY during which
the field was cropped. If you do not data from any previous year, you can
provide a year for labeling and annotations sake in output figures.}

\item{\code{CY_EXP}}{The year of interest of the selected experimental variable.
This is the year that the experimental variable was applied to grow the
crop in the selected 'CY_RESP' year. This is separate from the CY_RESP
variable because in some instances, applied data was categorized as or
applied in the year prior (i.e. WW seeding rates occur in fall of the year
before harvest in the following August).}

\item{\code{PY_EXP}}{The application year of the experimental variable that was
used to grow the crop in the 'PY_RESP' year. If you do not data from any
previous year, you can provide a year for labeling and annotations sake in
output figures.}

\item{\code{GRID}}{Determines the location of the aggregated data. Either select
'Grid' or 'Observed'. 'Grid' aggregates data to the centroids of a 10m grid
laid across the field, while 'Observed' aggregates data to the locations
of the observed response variable (yield or protein). Note that when
'Satellite' is selected there are no observed points to use, so the 'Grid'
option is selected by default.}

\item{\code{DAT_USED}}{Option for the length of year to use for CY data. In
winter wheat conventional systems and organic spring wheat systems,
there is a decision point around March 30th at which farmers must make
decisions on their fertilizer or seeding rates, respectively. When making
these decisions, farmers will only have data up to that decision point. This
options determines the covariate data aggregated for the CY, where
'Decision Point' aggregates data for the CY from January 1st to March 30th
of that year, while 'Full Year' aggregates data for the CY from January 1st
to December 31st. This is a residual special use case for Paul Hegedus' 2020
ICPA work.}

\item{\code{CY_RESP_FILES}}{Vector of names. Based on the user's selected field
and year, the database can be queried for the original file names of the
data uploaded to the database. The user selects the correct file name that
corresponds to the response variable data in the specified field and year.
In this case the 'CY_RESP' year. Multiple files are allowed for selection
because of the cases in which multiple files correspond to the data in the
given year and field (i.e. if harvest took multiple days etc.).}

\item{\code{PY_RESP_FILES}}{Vector of names. Based on the user's selected field
and year, the database can be queried for the original file names of the
data uploaded to the database. The user selects the correct file name that
corresponds to the response variable data in the specified field and year.
In this case the 'PY_RESP' year. Multiple files are allowed for selection
because of the cases in which multiple files correspond to the data in the
given year and field (i.e. if harvest took multiple days etc.). If you do
not have data available from a previous year and are not using the
interactive input selectors, simply input "None", otherwise when using the
interactive methods this is handled for you.}

\item{\code{CY_EXP_FILES}}{Data.frame with 2 columns, 'orig_file' and 'table'.
Based on the user's selected field and year, the database can be queried
for the original file names of the data uploaded to the database. The user
selects the correct file name that corresponds to the experimental variable
data in the specified field and year. In this case, the 'CY_EXP' year.
Multiple files are allowed for selection because of the cases in which
multiple files correspond to the data in the given year and field (i.e.
if application took multiple days etc.). The 'orig_file' column contains
the name of the original filename uploaded to the database with the raw
data and 'table' contains the table within the farmer schema that the
data is housed. This is because some as-applied experimental data are
point vectors and some are polygons.}

\item{\code{PY_EXP_FILES}}{Data.frame with 2 columns, 'orig_file' and 'table'.
Based on the user's selected field and year, the database can be queried
for the original file names of the data uploaded to the database. The user
selects the correct file name that corresponds to the experimental variable
data in the specified field and year. In this case, the 'PY_EXP' year.
Multiple files are allowed for selection because of the cases in which
multiple files correspond to the data in the given year and field (i.e. if
application took multiple days etc.). The 'orig_file' column contains
the name of the original filename uploaded to the database with the raw
data and 'table' contains the table within the farmer schema that the
data is housed. This is because some as-applied experimental data are
point vectors and some are polygons. If no data from the desired year
is available, put "None" in the 'orig_file' column.}

\item{\code{saveInDB}}{Yes/No. Logical, whether to save the aggregated data into
the OFPE formatted database. Not an option if you are supplying your own
boundary for which to aggregate data.}

\item{\code{export}}{Yes/No. Logical, whether to export the aggregated data as
a '.csv' file. If yes, the user will need to provide the 'exportName'.}

\item{\code{exportName}}{If exporting the aggregated data as a '.csv', the user
needs to specify the name of the file to export. This includes the file
path.}

\item{\code{CY_RESP_COL}}{Data.frame with 3 columns, 'RESP', 'DIST', and
'orig_file'. This data.frame contains the column containing data for
the selected response variable ('RESP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional (can be NA) and optionally used during a cleaning
step where points are removed if they are more than 4SD from the mean
distance between observations as this indicates that the equipment was
moving at an abnormal speed and potentially resulting in erroneous
measurements. The 'orig_file' is the same as those selected in the
'CY_RESP_FILES' vector.}

\item{\code{PY_RESP_COL}}{Data.frame with 3 columns, 'RESP', 'DIST', and
'orig_file'. This data.frame contains the column containing data for
the selected response variable ('RESP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional and optionally used during a cleaning step where
points are removed if they are more than 4SD from the mean distance
between observations as this indicates that the equipment was moving
at an abnormal speed and potentially resulting in erroneous measurements.
The 'orig_file' is the same as those selected in the 'PY_RESP_FILES' vector.}

\item{\code{CY_EXP_COL}}{Data.frame with 4 columns, 'EXP', 'DIST', 'PRODUCT',
and 'orig_file'. This data.frame contains the column containing data for
the selected experimental variable ('EXP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional (can be NA) and optionally used during a cleaning
step where points are removed if they are more than 4SD from the mean
distance between observations as this indicates that the equipment was
moving at an abnormal speed and potentially resulting in erroneous
measurements. The 'PRODUCT' column is also optional, and not relevant
for seeding rate data, and used if there is a column corresponding to
the product applied. When left blank, it is assumed that the 'EXP' column
contains data in lbs/acre and the user is not given an option to provide
a conversion rate. It is good practice to always select a column (even if
no specific 'product' column) here to explicitly state a conversion factor.
The 'orig_file' is the same as those selected in the 'CY_EXP_FILES' table.}

\item{\code{PY_EXP_COL}}{Data.frame with 4 columns, 'EXP', 'DIST', 'PRODUCT',
and 'orig_file'. This data.frame contains the column containing data for
the selected experimental variable ('EXP') and the column (if any) that
correspond to the distance between observations ('DIST'). The 'DIST'
selection is optional (can be NA) and optionally used during a cleaning
step where points are removed if they are more than 4SD from the mean
distance between observations as this indicates that the equipment was
moving at an abnormal speed and potentially resulting in erroneous
measurements. The 'PRODUCT' column is also optional, and not relevant
for seeding rate data, and used if there is a column corresponding to
the product applied. When left blank, it is assumed that the 'EXP' column
contains data in lbs/acre and the user is not given an option to provide
a conversion rate. It is good practice to always select a column (even if
no specific 'product' column) here to explicitly state a conversion factor.
The 'orig_file' is the same as those selected in the 'PY_EXP_FILES' table.}

\item{\code{CY_EXP_CONV}}{Data.frame with 3 columns, 'FORMULA', 'CONVERSION',
and 'orig_file'. Based on the users selection of 'PRODUCT' in the
CY_EXP_COL data.frame, the formula from the 'PRODUCT' column is extracted
and used to ask the user the desired conversion factor from the product
applied to lbs per acre. Again, this is only really applicable for fertilzier
rates unless seeding rates are reported in units besides lbs per acre. The
'orig_file' is the same as those selected in the 'CY_EXP_FILES' table.}

\item{\code{PY_EXP_CONV}}{Data.frame with 3 columns, 'FORMULA', 'CONVERSION',
and 'orig_file'. Based on the users selection of 'PRODUCT' in the
CY_EXP_COL data.frame, the formula from the 'PRODUCT' column is extracted
and used to ask the user the desired conversion factor from the product
applied to lbs per acre. Again, this is only really applicable for fertilzier
rates unless seeding rates are reported in units besides lbs per acre. The
'orig_file' is the same as those selected in the 'CY_EXP_FILES' table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'AggInputs' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectInputs"></a>}}
\if{latex}{\out{\hypertarget{method-selectInputs}{}}}
\subsection{Method \code{selectInputs()}}{
Interactive method for selecting aggregation input options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggInputs$selectInputs()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because passed in during class
instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A completed 'AggInputs' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AggInputs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
