% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{applyExpRates}
\alias{applyExpRates}
\title{Apply experimental rates across the room.}
\usage{
applyExpRates(rx_sdt, exp_rates, exp_rates_prop, fld_prop, exp_rate_length)
}
\arguments{
\item{rx_sdt}{A 'sf' object containing the grid cells available for
applying rates to.}

\item{exp_rates}{Provide a vector of experimental rates equal to the
number of experimental rates provided in 'exp_rate_length'. This is
required for all new experiments, however can be left to null for
experimental prescriptions if experimental rates should be generated
based on gaps in optimum rates.}

\item{exp_rates_prop}{Provide proportions (0 - 1) for the length
of experimental rates provided in 'exp_rate_length'. This is required
for all new experiments and experimental prescriptions.}

\item{fld_prop}{The proportion of the field to apply experimental
rates to (i.e. 0.5 or 1.0), OR, the percent of available cells to
apply check rates to (i.e. 0.05, 0.1).}

\item{exp_rate_length}{Provide the length of experimental rates to apply.
This applies to new experiments and experimental prescriptions. This
represents the equipment constraints of the farmer. In the case of the
experimental prescription, this number of rates does not include the
number of rates for the optimized base map, so take your selections for
the management scenario and number of optimum rates into account.}

\item{strat_dat}{List of data to stratify data on from the OFPE
database... TODO... not implemented yet}
}
\value{
Amended 'rx_sdt' object wit randomly placed experimental
rates applied.
}
\description{
This function applies experimental rates across a field
randomly, or stratified on user supplied data. The arguments contain
an 'sf' object with the grid cells available for applying rates to.
Based on the experimental rates and associated proportions.
}
