% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesLawrence.R
\name{BayesLawrence}
\alias{BayesLawrence}
\title{R6 Class for Bayesian modeling of crop responses using modified model from Lawrence et al. 2015}
\description{
R6 class using Bayesian Non-Linear Regression to fit a crop
response model with the experimental variable, precipitation, and claycontent.
This is a modified version of the model developed by Patrick Lawrence (2015) in
"A probabilistic Bayesian framework for progressively updating site-specific recommendations".
Rather than using EC data due to its unavailability, clay content is used because of
its relation to EC. This class is initialized with a named list of training (named 'trn')
and validation (named 'val') datasets, the response variable, the experimental
variable, and the means of the centered data.

Any covariates passed into the model are replaced by precipitation and claycontent.
This is used to fit the final model that is returned to the user for use in the
simulation to predict the response under varying rates of the experimental variable.

The 'saveDiagnostics' method include residuals vs. fitted. The fitting process
also prepares data for validation plots in the
'ModClass' R6 class. This includes predicting observations in the validation
dataset, making a unique id using the year and fieldname, uncentering data, and
identifying a field name to use for plotting that reflects all fields in the
dataset.
}
\seealso{
\code{\link{ModClass}} for the class that calls the ModClass interface,
\code{\link{NonLinear_Logistic}}, \code{\link{GAM}}, and \code{\link{RF}}
for alternative model classes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dat}}{Named list of traning (named 'trn') and validation (named 'val')
datasets with the response, experimental, and remotely sensed variables.}

\item{\code{respvar}}{Character, the response variable of interest.}

\item{\code{expvar}}{Character, the experimental variable of interest.}

\item{\code{covars}}{Character vector of covariates to use for training the model. Ignored,
because only the experimental variable, precipitation, and claycontent are used.}

\item{\code{m}}{Fitted RF model.}

\item{\code{form}}{Final RF formula.}

\item{\code{parm_df}}{Data frame of parameter names, and a column
named 'bad_parms' to indicate whether to include in the model formula. Also includes
columns for the mean and standard deviation of each parameter.}

\item{\code{fieldname}}{Unique name for the field(s) analyzed. If multiple fields are used
they are separated by an ampersand, otherwise the singular field name is used. This
is used for plotting.}

\item{\code{mod_type}}{Name of the model of this class, used for plotting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BayesLawrence$new()}}
\item \href{#method-fitMod}{\code{BayesLawrence$fitMod()}}
\item \href{#method-predResps}{\code{BayesLawrence$predResps()}}
\item \href{#method-saveDiagnostics}{\code{BayesLawrence$saveDiagnostics()}}
\item \href{#method-clone}{\code{BayesLawrence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
The initialization creates a data frame ('parm_df') containing the parameter names,
the mean and standard deviation, and whether it meets
the criteria to be omitted from the model, making it a 'bad_parm'. The criteria for
this is over 30\% of data for a given year missing for a parameter or a standard
deviation of zero, indicating singularity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLawrence$new(dat, respvar, expvar, covars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Named list of training (named 'trn') and validation (named 'val')
datasets with the response, experimental, and remotely sensed variables.}

\item{\code{respvar}}{Character, the response variable of interest.}

\item{\code{expvar}}{Character, the experimental variable of interest.}

\item{\code{covars}}{covars Character vector of covariates to use for training the model. Ignored,
because only the experimental variable, precipitation, and claycontent are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A instantiated 'RF' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fitMod"></a>}}
\if{latex}{\out{\hypertarget{method-fitMod}{}}}
\subsection{Method \code{fitMod()}}{
Method for fitting the modified Lawrence et al. 2015 Bayesian model to
response variables using experimental input data, precipitation, and claycontent.

A model is fit that is returned  to the user for use in the simulation to
predict the response under varying  rates of the experimental variable.

Finally, this method prepares the validation data for plotting by using the model to predict
the response for each of the observations in the validation dataset, uncentering data if
necessary, and identifying a unique field name from the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLawrence$fitMod()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{Parameters provided upon class instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A fitted BayesLawrence model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predResps"></a>}}
\if{latex}{\out{\hypertarget{method-predResps}{}}}
\subsection{Method \code{predResps()}}{
Method for predicting response variables using data and a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLawrence$predResps(dat, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data for predicting response variables for.}

\item{\code{m}}{The fitted model to use for predicting the response
variable for each observation in 'dat'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of predicted values for each location in 'dat'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDiagnostics"></a>}}
\if{latex}{\out{\hypertarget{method-saveDiagnostics}{}}}
\subsection{Method \code{saveDiagnostics()}}{
Method for saving diagnostic plots of the fitted model. This is
simply the residuals vs fitted values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLawrence$saveDiagnostics(out_path, SAVE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the model fitting process}

\item{\code{SAVE}}{Whether to save diagnostic plots.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Diagnostic plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLawrence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
