% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesLinear.R
\name{BayesLinear}
\alias{BayesLinear}
\title{R6 Class for Bayesian Linear modeling of crop responses}
\description{
R6 class using Bayesian Linear Regression to fit a crop
response model with the experimental variable and remotely sensed covariate
data. This class is initialized with a named list of training (named 'trn')
and validation (named 'val') datasets, the response variable, the experimental
variable, and the means of the centered data.

The initialization creates a data frame ('parm_df') containing the parameter names,
the mean and standard deviation, and whether it meets
the criteria to be omitted from the model, making it a 'bad_parm'. The criteria for
this is over 30\% of data for a given year missing for a parameter or a standard
deviation of zero, indicating singularity.

The process then creates a formula for a final model with all parameters that
are not considered 'bad'. This is used to fit the final model that is returned
to the user for use in the simulation to predict the response under varying
rates of the experimental variable.

The 'saveDiagnostics' method include residuals vs. fitted. The fitting process
also prepares data for validation plots in the
'ModClass' R6 class. This includes predicting observations in the validation
dataset, making a unique id using the year and fieldname, uncentering data, and
identifying a field name to use for plotting that reflects all fields in the
dataset.
}
\seealso{
\code{\link{ModClass}} for the class that calls the ModClass interface,
\code{\link{NonLinear_Logistic}}, \code{\link{GAM}}, and \code{\link{RF}}
for alternative model classes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dat}}{Named list of traning (named 'trn') and validation (named 'val')
datasets with the response, experimental, and remotely sensed variables.}

\item{\code{respvar}}{Character, the response variable of interest.}

\item{\code{expvar}}{Character, the experimental variable of interest.}

\item{\code{covars}}{Character vector of covariates to use for training the model.}

\item{\code{m}}{Fitted RF model.}

\item{\code{form}}{Final RF formula.}

\item{\code{parm_df}}{Data frame of parameter names, and a column
named 'bad_parms' to indicate whether to include in the model formula. Also includes
columns for the mean and standard deviation of each parameter.}

\item{\code{fieldname}}{Unique name for the field(s) analyzed. If multiple fields are used
they are separated by an ampersand, otherwise the singular field name is used. This
is used for plotting.}

\item{\code{mod_type}}{Name of the model of this class, used for plotting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BayesLinear-new}{\code{BayesLinear$new()}}
\item \href{#method-BayesLinear-fitMod}{\code{BayesLinear$fitMod()}}
\item \href{#method-BayesLinear-predResps}{\code{BayesLinear$predResps()}}
\item \href{#method-BayesLinear-saveDiagnostics}{\code{BayesLinear$saveDiagnostics()}}
\item \href{#method-BayesLinear-clone}{\code{BayesLinear$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BayesLinear-new"></a>}}
\if{latex}{\out{\hypertarget{method-BayesLinear-new}{}}}
\subsection{Method \code{new()}}{
The initialization creates a data frame ('parm_df') containing the parameter names,
the mean and standard deviation, and whether it meets
the criteria to be omitted from the model, making it a 'bad_parm'. The criteria for
this is over 30\% of data for a given year missing for a parameter or a standard
deviation of zero, indicating singularity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLinear$new(dat, respvar, expvar, covars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Named list of training (named 'trn') and validation (named 'val')
datasets with the response, experimental, and remotely sensed variables.}

\item{\code{respvar}}{Character, the response variable of interest.}

\item{\code{expvar}}{Character, the experimental variable of interest.}

\item{\code{covars}}{Character vector of covariates to use for training the model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A instantiated 'RF' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BayesLinear-fitMod"></a>}}
\if{latex}{\out{\hypertarget{method-BayesLinear-fitMod}{}}}
\subsection{Method \code{fitMod()}}{
Method for fitting the Bayesian linear model to response variables using
experimental and covariate data.

The fitting begins by taking the data frame ('parm_df') containing the parameter names,
the mean and standard deviation, and identifying whether
each parameter meets the criteria to be omitted from the model, making it a 'bad_parm'.
The criteria for this is over 30\% of data for a given year missing for a parameter or a
standard deviation of zero, indicating singularity.

OLD: A subset of 1000 observations is taken to perform top-down feature
selection based on the RMSE from predictions to a holdout set. If a covariate
lowers the RMSE when removed, it is flagged as 'bad'.
NEW: No feature selection is performed due to computing time.

The process then creates a formula for a final model with all parameters that
are not considered 'bad'. This is used to fit the final model that is returned
to the user for use in the simulation to predict the response under varying
rates of the experimental variable.

Finally, this method prepares the validation data for plotting by using the model to predict
the response for each of the observations in the validation dataset, uncentering data if
necessary, and identifying a unique field name from the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLinear$fitMod()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{Parameters provided upon class instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A fitted BayesLinear model.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BayesLinear-predResps"></a>}}
\if{latex}{\out{\hypertarget{method-BayesLinear-predResps}{}}}
\subsection{Method \code{predResps()}}{
Method for predicting response variables using data and a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLinear$predResps(dat, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data for predicting response variables for.}

\item{\code{m}}{The fitted model to use for predicting the response
variable for each observation in 'dat'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of predicted values for each location in 'dat'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BayesLinear-saveDiagnostics"></a>}}
\if{latex}{\out{\hypertarget{method-BayesLinear-saveDiagnostics}{}}}
\subsection{Method \code{saveDiagnostics()}}{
Method for saving diagnostic plots of the fitted model. This is
simply the residuals vs fitted values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLinear$saveDiagnostics(out_path, SAVE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the model fitting process}

\item{\code{SAVE}}{Whether to save diagnostic plots.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Diagnostic plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BayesLinear-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BayesLinear-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLinear$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
