% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Billings_PremDock_2016.R
\docType{data}
\name{Billings_PremDock_2016}
\alias{Billings_PremDock_2016}
\title{Default protein premium/dockage data used for OFPE simulations.}
\format{
A data frame with 33 rows and 2 variables:
\describe{
\item{pro}{Protein percent level (real number).}
\item{PremDock}{Price gained or lost from the base price related to the associated protein level, in dollars.}
}
}
\usage{
Billings_PremDock_2016
}
\description{
A dataset containing protein levels of hard red winter wheat and
the associated premium or dockage gathered in 2016 from the Billings,
MT grain elevator. This dataset is used to fit a function of premium
dockage related to grain protein that is used in the OFPE Monte Carlo
simulation to calculate estimated net-return.
}
\keyword{datasets}
