% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDB.R
\name{BuildDB}
\alias{BuildDB}
\title{R6 Class for building OFPE database}
\source{
\url{https://trac.osgeo.org/postgis/wiki/UsersWikiCreateFishnet}
}
\description{
R6 Class for building OFPE database

R6 Class for building OFPE database
}
\details{
R6 class for for loading required extensions and
building the the OFPE database schemas and initial tables.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db}}{Database connection, such as PostgreSQLConnection.}

\item{\code{postgis_version}}{Version of PostGIS installed (i.e. 2.5, 3.0).}

\item{\code{farmers}}{Vector of farmer names to use for building initial schemas.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BuildDB$new()}}
\item \href{#method-buildDatabase}{\code{BuildDB$buildDatabase()}}
\item \href{#method-clone}{\code{BuildDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a database builder object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildDB$new(db = NA, postgis_version = NA, farmers = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Database connection.}

\item{\code{postgis_version}}{Version of PostGIS.}

\item{\code{farmers}}{Vector of farmer names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{BuildDB} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildDatabase"></a>}}
\if{latex}{\out{\hypertarget{method-buildDatabase}{}}}
\subsection{Method \code{buildDatabase()}}{
Execute database builder functions.
@description
Loads extensions needed for OFPE database functions such
as PostGIS tools and a function for generating a net across an area of
interest (see source).
@param db Connection to a database.
@param postgis_version character PostGIS version installed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildDB$buildDatabase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BuildDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
