% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUTMzone.R
\name{calcUTMzone}
\alias{calcUTMzone}
\title{Calculate the Universal Transverse Mercator coordinate system zone}
\source{
https://apollomapping.com/blog/gtm-finding-a-utm-zone-number-easily
}
\usage{
calcUTMzone(FILE)
}
\arguments{
\item{FILE}{Spatial shapefile for determining UTM zone.}
}
\value{
WGS84 UTM code (i.e. 32612, 32614).
}
\description{
Calculates the UTM zone for the specified data and
returns a UTM code for WGS84 (i.e. 32612, 32614). This function is
performed in the ManageParms class to store the UTM zone for the farm.
A shapefile is passed in with the 'FILE' argument, the spatial data is
used to determine the UTM zone mathematically, same if the 'bounds'
argument is used. The returned value is the UTM EPSG code for the WGS84 datum.
}
