% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{convPolyToMulti}
\alias{convPolyToMulti}
\title{Convert polygon table geometry}
\usage{
convPolyToMulti(db, dat, schema, dtype)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{dat}{Data of 'sp' object that was imported to the
specified schema and table to convert to multipolygon
geometry. Must have a column named 'geometry'.}

\item{schema}{Name of the schema with the table to convert
geometry in.}

\item{dtype}{The name of the table to convert geometry in.}
}
\value{
Polygon geometry type converted to MultiPolygon.
}
\description{
Convert a table with a Polygon geometry to a
table with a MultiPolygon geometry.
}
