% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendMap.R
\name{create_orientation_arrow}
\alias{create_orientation_arrow}
\title{Create orientation arrow.}
\source{
https://github.com/3wen/legendMap.git
}
\usage{
create_orientation_arrow(scale_bar, length, distance = 1, dist_units = "km")
}
\arguments{
\item{scale_bar}{Result of create_scale_bar()}

\item{length}{Desired length of the arrow}

\item{distance}{Distance between legend rectangles and the bottom of the arrow.
Default is 1.}

\item{dist_units}{Units of distance "km" (kilometers) (default), "nm"
(nautical miles), "mi" (statute miles).}
}
\value{
Return a list whose elements are; 'res': coordinates to draw an arrow,
and 'coords_n': coordinates of the middle of the arrow (where the "N" will be plotted).
}
\description{
Function from 3wen/legendMap repo for 'adding a North arrow and/or a
scale bar to maps done using ggplot2'. All credit for this function goes to
Ewen Gallic. All notation and documentation is from Ewen Gallic.

This function creates an orientation arrow (i.e. north arrow).
See https://github.com/3wen/legendMap.git for examples.
}
