% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatInputs.R
\name{DatInputs}
\alias{DatInputs}
\title{R6 Class for storing inputs to the 'DatObject' class}
\description{
R6 class for for storing information needed for the
'DatObject' class which is used for the analysis/simulation and Rx
building steps of the OFPE data cycle. This object inlcudes user
selections such as the field and year of data to export from an
OFPE database and the type of data (grid or observed) for analysis
and simulation/prescription generation.

Inputs can be supplied directly to this class during instantiation, however
this is NOT recommended except for advanced users. It is recommended that the
user supplies the database connection and uses the interactive selection
methods to select user inputs.

This class is passed to the 'DatObject' class that executes the methods
for exporting data from the database and processing the data for analysis,
simulation, and prescription building.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{farmername}}{Name of the farmer that owns the selected field.}

\item{\code{fieldname}}{Name of the field to aggregate data for. Selected from
the 'all_farms.fields' table of an OFPE formatted database.}

\item{\code{respvar}}{Response variable to aggregate data for, select/input
'Yield' or 'Protein'. Multiple options allowed.}

\item{\code{expvar}}{Experimental variable to aggregate data for, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{yldyears}}{The year(s) of interest for the yield response
variables in the selected field. This must be a named list with the
specified field names.}

\item{\code{proyears}}{The year(s) of interest for the protein response
variables in the selected field. This must be a named list with the
specified field names.}

\item{\code{analysis_grid}}{Select whether to use gridded or observed data
locations for the analysis step. See the 'AggInputs' class for more
information on the 'GRID' option. The user must have aggregated data
with the specified GRID option prior to this step. (i.e. you will not
have access to data aggregated with the 'Grid' option if you have not
executed the process of aggregation with the 'Grid' option selected. The
same principle applies for the 'Observed' option. It is recommended that
the analysis is performed with 'Observed' data, and for the simulation to
be performed with 'Grid' data.}

\item{\code{sim_grid}}{Select whether to use gridded or observed data
locations for the simulation and subsequent prescription building step.
See the 'AggInputs' class for more information on the 'GRID' option.
The user must have aggregated data with the specified GRID option prior
to this step. (i.e. you will not have access to data aggregated with the
'Grid' option if you have not executed the process of aggregation with
the 'Grid' option selected. The same principle applies for the 'Observed'
option. It is recommended that the analysis is performed with 'Observed'
data, and for the simulation tobe performed with 'Grid' data.}

\item{\code{dat_used}}{Option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection.}

\item{\code{veg_index}}{Option for the vegetation index data to use for analysis,
simulation, and prescription building steps. Select from 'NDVI', 'NDRE',
and 'CIRE'. These are the three vegatation indices downloaded from
Google Earth Engine and used to enrich yield and protein datasets in the
aggregation step.}

\item{\code{prec_source}}{Option for the satellite source of precipitation data
to use for analysis, simulation, and prescription building steps. Select
from 'Daymet' or 'Gridmet'. These are the sources of precipitation data
downloaded from Google Earth Engine and used to enrich yield and protein
datasets in the aggregation step. This selection is for the users preference,
but note that the other source will be used if the selected source is
unavailable.}

\item{\code{gdd_source}}{Option for the satellite source of growing degree day (GDD)
data to use for analysis, simulation, and prescription building steps. Select
from 'Daymet' or 'Gridmet'. These are the sources of GDD data downloaded from
Google Earth Engine and used to enrich yield and protein datasets in the
aggregation step. This selection is for the users preference, but note that
the other source will be used if the selected source is unavailable.}

\item{\code{center}}{TRUE/FALSE. Option for whether to center explanatory data around
each explanatory variables mean or to use the raw observed explanatory varaible
data. Centering is recommended as it puts variables on similar scales and makes
the model fitting process less error prone.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DatInputs$new()}}
\item \href{#method-selectInputs}{\code{DatInputs$selectInputs()}}
\item \href{#method-clone}{\code{DatInputs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatInputs$new(
  dbCon,
  farmername = NULL,
  fieldname = NULL,
  respvar = NULL,
  expvar = NULL,
  yldyears = NULL,
  proyears = NULL,
  analysis_grid = NULL,
  sim_grid = NULL,
  dat_used = NULL,
  veg_index = NULL,
  prec_source = NULL,
  gdd_source = NULL,
  center = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{farmername}}{Name of the farmer that owns the selected field.}

\item{\code{fieldname}}{Name of the field to aggregate data for. Selected from
the 'all_farms.fields' table of an OFPE formatted database.}

\item{\code{respvar}}{Response variable to aggregate data for, select/input
'Yield' or 'Protein'. Multiple options allowed.}

\item{\code{expvar}}{Experimental variable to aggregate data for, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{yldyears}}{The year(s) of interest for the yield response
variables in the selected field. This must be a named list with the
specified field names.}

\item{\code{proyears}}{The year(s) of interest for the protein response
variables in the selected field. This must be a named list with the
specified field names.}

\item{\code{analysis_grid}}{Select whether to use gridded or observed data
locations for the analysis step. See the 'AggInputs' class for more
information on the 'GRID' option. The user must have aggregated data
with the specified GRID option prior to this step. (i.e. you will not
have access to data aggregated with the 'Grid' option if you have not
executed the process of aggregation with the 'Grid' option selected. The
same principle applies for the 'Observed' option. It is recommended that
the analysis is performed with 'Observed' data, and for the simulation to
be performed with 'Grid' data.}

\item{\code{sim_grid}}{Select whether to use gridded or observed data
locations for the simulation and subsequent prescription building step.
See the 'AggInputs' class for more information on the 'GRID' option.
The user must have aggregated data with the specified GRID option prior
to this step. (i.e. you will not have access to data aggregated with the
'Grid' option if you have not executed the process of aggregation with
the 'Grid' option selected. The same principle applies for the 'Observed'
option. It is recommended that the analysis is performed with 'Observed'
data, and for the simulation tobe performed with 'Grid' data.}

\item{\code{dat_used}}{Option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection.}

\item{\code{veg_index}}{Option for the vegetation index data to use for analysis,
simulation, and prescription building steps. Select from 'NDVI', 'NDRE',
and 'CIRE'. These are the three vegatation indices downloaded from
Google Earth Engine and used to enrich yield and protein datasets in the
aggregation step.}

\item{\code{prec_source}}{Option for the satellite source of precipitation data
to use for analysis, simulation, and prescription building steps. Select
from 'Daymet' or 'Gridmet'. These are the sources of precipitation data
downloaded from Google Earth Engine and used to enrich yield and protein
datasets in the aggregation step. This selection is for the users preference,
but note that the other source will be used if the selected source is
unavailable.}

\item{\code{gdd_source}}{Option for the satellite source of growing degree day (GDD)
data to use for analysis, simulation, and prescription building steps. Select
from 'Daymet' or 'Gridmet'. These are the sources of GDD data downloaded from
Google Earth Engine and used to enrich yield and protein datasets in the
aggregation step. This selection is for the users preference, but note that
the other source will be used if the selected source is unavailable.}

\item{\code{center}}{TRUE/FALSE. Option for whether to center explanatory data around
each explanatory variables mean or to use the raw observed explanatory varaible
data. Centering is recommended as it puts variables on similar scales and makes
the model fitting process less error prone.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'AggInputs' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-selectInputs"></a>}}
\if{latex}{\out{\hypertarget{method-selectInputs}{}}}
\subsection{Method \code{selectInputs()}}{
Interactive method for selecting inputs related to the data used in the
analysis, simulation, and subsequent prescription generation steps. The
description below describes the process of interactively selecting the
necessary parameters needed for the automated analysis, simulation, and
prescription building.

The user first selects a farmer for which they want to analyze a field
from, which is used to compile a list of available fields ready for
analysis, indicated by its presence in the farmername_a schema of the
OFPE database.

The user then selects the response variables to optimize on and the
experimental variable to optimize. The user must know what data is
available for the specific field (i.e. if the user select 'Protein' they
must have aggregated protein data for the specified field, or if the
user selects 'As-Applied Seed Rate' seed rates must have been the
experimental variable of interest when aggregating data).

The user then selects the location of aggregated data to use for both
the analysis and simulation/prescription building steps. The user also
needs to select the length of the year for which 'current' year data
was aggregated for (March 30th decision point or the full year).

The user also has the choice of which vegetation index data to use as
covariates, as well as the preferred source for precipitation and
growing degree day data. Finally, the user has the option of whether
to center covariate data or to use the raw observed data for analysis
and simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatInputs$selectInputs()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because passed in during class
instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A completed 'DatInputs' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatInputs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
