% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBCon.R
\name{DBCon}
\alias{DBCon}
\title{R6 Class for connecting to a database.}
\description{
R6 class for for connecting to your OFPE
database. The default database driver is PostgreSQL as that
is recommended for OFPE use and shown in the OFPE tutorials and
vignettes. However, any database driver can be supplied. This class
was created to consolidate database connections being supplied to
functions and other R6 classes in the OFPE data workflow.

Users can connect with any DBI compliant option. Using RPostgreSQL drivers on
a Mac, users can provide the user, password, dbname, host, and port. If on a
Windows computer the user just needs to set up a PostgreSQL ODBC driver, and
supply the DSN, user, and password.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db}}{Database connection, such as PostgreSQL connection.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DBCon$new()}}
\item \href{#method-disconnect}{\code{DBCon$disconnect()}}
\item \href{#method-clone}{\code{DBCon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a database connection inside of the DBCon class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBCon$new(
  user = NA,
  password = NA,
  dbname = NULL,
  host = NULL,
  port = NULL,
  dsn = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user}}{Username associated with the database owner.}

\item{\code{password}}{Password associated with the database owner.}

\item{\code{dbname}}{Name of the database to connect to.}

\item{\code{host}}{Host name of the database location (typically localhost)
unless user is connecting to a non-local database.}

\item{\code{port}}{The port number of the database.}

\item{\code{dsn}}{Optional. If ODBC database source name is setup. Required for
Windows users.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An open database connection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Close a database connection in the DCCon class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBCon$disconnect()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No parameters necessary. Simply call the method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A closed database connection.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBCon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
