% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derivFFoptCpp}
\alias{derivFFoptCpp}
\title{Function for identifying the optimum full-field experimental rate based on
profit maximization. This function selects the optimum rate as
the rate where the difference in full-field net-return between
experimental rates (first derivitave of the net-return vs. experimental rate
function) does not exceed the cost of applying one more unit of
the experimental rate across the entire field.}
\usage{
derivFFoptCpp(NRff, rr, fieldsize, CEXP)
}
\arguments{
\item{NRff}{A matrix with rows for each experimental rate and
columns for the rate and the full-field net-return calculated
under a uniform application of the corresponding rate.}

\item{rr}{Integer of the number of observations (rows of NRdf).}

\item{fieldsize}{Double, the size of the field(s) that the
experimental input would be applied to.}

\item{CEXP}{Double, the cost of the experimental input.}
}
\value{
NRffmax, table with the full-field optimum rate
and the full-field net-return.
}
\description{
This function calculates the optimum rate from the sum of the
net-return for a uniform rate applied across the field to
identify the full-field optimum rate.
}
\details{
See the documentation in OFPE::derivNRoptCpp for a site-specific
example of how the optimum is identified compared to when the
optimization method is set to 'max'.
}
