% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpGen.R
\name{ExpGen}
\alias{ExpGen}
\title{R6 Class for generating a OFPE experiments}
\description{
R6 class for for creating a prescription or experiment for a
field of interest. The user can create a new experiment with inputs randomly
applied across the field with no stratification, or select data
on which to stratify experimental rates. This class randomly places experimental
input rates across a field(s) selected from the database, with the user's
choice for stratification.

This class follows the generator interface that includes an initialization method
and an 'executeOutput' method.
}
\seealso{
\code{\link{DBCon}} for the database connection class, and
\code{\link{RxGen}} for the alternative class that creates
experimental prescriptions or pure prescriptions.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{trt_length}}{Length, in meters, for which to apply treatments.}

\item{\code{boom_width}}{The width of the sprayer boom or spreader.}

\item{\code{orientation}}{TODO... Not implemented yet.}

\item{\code{fld_prop}}{The proportion of the field to apply experimental
or optimum check rates to.}

\item{\code{expvar}}{Experimental variable to optimize, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that is experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{conv}}{The conversion factor between lbs of the input to the
units of the as-applied input (i.e. lbs N/ac to unit input/ac)}

\item{\code{base_rate}}{The rate to apply between the experimental rates
and the field edge, or as check rates in the prescription selected
option.}

\item{\code{rx_for_year}}{Provide the year that the experiment
is made for. Used for labeling outputs.}

\item{\code{SAVE}}{Logical, whether to save figures and the experiment.
Autofilled to FALSE if a user selects NA in the 'out_path' or is NULL.
Autofilled to TRUE otherwise.}

\item{\code{out_path}}{Provide the path to the folder in which to store and
save figures and the prescription Type NA to not create any folders.
You will not be able to save any outputs. (Note, even if a path is provided,
the user can pass FALSE as the sole argument to the 'setupOP' method
to prevent the creation of folders. This will automatically prevent
any plots to be saved.).}

\item{\code{farmername}}{If the user is creating a new experiment, provide or
select the name of the farmer that owns or manages the field(s) that
an experiment is going to be generated for.}

\item{\code{fieldname}}{If the user is creating a new experiment, provide or
select the field names of the field to use. The field list is from the
available fields in the database for experimentation.}

\item{\code{exp_rate_length}}{Provide the length of experimental rates to apply.
This applies to new experiments and experimental prescriptions. This
represents the equipment constraints of the farmer. In the case of the
experimental prescription, this number of rates does not include the
number of rates for the optimized base map, so take your selections for
the management scenario and number of optimum rates into account.}

\item{\code{exp_rates}}{Provide a vector of experimental rates equal to the
number of experimental rates provided in 'exp_rate_length'. This is
required for all new experiments, however can be left to null for
experimental prescriptions if experimental rates should be generated
based on gaps in optimum rates.}

\item{\code{exp_rates_prop}}{Provide proportions (0 - 1) for the length
of experimental rates provided in 'exp_rate_length'. This is required
for all new experiments and experimental prescriptions.}

\item{\code{strat_dat}}{If the user is creating a new experiment with stratification
they must apply the data that they are stratifying on. This provides an
identifier for the data being used for the stratification. These must be codes
used in the OFPE workflow such as 'yld', 'pro', 'aa_n', 'aa_sr', etc.}

\item{\code{strat_dat_year}}{Follows the same structure of 'strat_dat', however contains
the years to use for stratification of each of the data types for each of the fields.}

\item{\code{unique_fieldname}}{Unique fieldname for the field(s) used for the experiment. This
concatenates multiple fields with an ampersand. Used for labeling.}

\item{\code{rx_dt}}{Data frame that contains the coordinates of locations to apply
experimental inputs. For an experiment, these are the centroids of the grid
made to aggregate data for the field.}

\item{\code{RX}}{Table containing the geographic locations of the experiment. Also
contains the rates in the experimental and as-applied units. This can be
saved as a shapefile and given to the equipment applying the input.}

\item{\code{out_name}}{Created parameter including the unique fieldname and year the
experiment is being generated for.}

\item{\code{out_map_name}}{Created parameter with the file name for the map of the
prescription. Used for saving the map to the 'Outputs' folder.}

\item{\code{cell_out_map_name}}{Created parameter with the file name for the map of the
rate type applied to each cell. Used for saving the map to the 'Outputs' folder.}

\item{\code{var}}{The label of the variable to map. Used in figure labeling for plotting
in RxClass.}

\item{\code{var_col_name}}{The name of the column of the variable in the
supplied data ('dat'). Used in figure labeling for plotting
in RxClass.}

\item{\code{var_label}}{The label to be applied to the legend of the map
corresponding to the variable mapped. Used in figure labeling for plotting
in RxClass.}

\item{\code{var_main_label}}{The main label to apply to the map. Used in figure
labeling for plotting in RxClass.}

\item{\code{mgmt_scen}}{For this class, the management scenario is always an experiment
so this parameter is set to 'exp'.}

\item{\code{size}}{The size of the treatment zones, which is the treatment length x
the boom width x 2.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ExpGen$new()}}
\item \href{#method-executeOutput}{\code{ExpGen$executeOutput()}}
\item \href{#method-clone}{\code{ExpGen$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It is recommended to initialize this class through the RxClass,
because it is integrated into the OFPE workflow. This also ensures all of the
inputs are in the correct format and present.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpGen$new(
  dbCon,
  trt_length,
  boom_width,
  orientation,
  fld_prop,
  expvar,
  conv,
  base_rate,
  rx_for_year,
  out_path = NULL,
  SAVE,
  fieldname,
  farmername,
  exp_rate_length,
  exp_rates,
  exp_rates_prop,
  strat_dat = NULL,
  strat_dat_year = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{trt_length}}{Length, in meters, for which to apply treatments.}

\item{\code{boom_width}}{The width of the sprayer boom or spreader.}

\item{\code{orientation}}{TODO... Not implemented yet.}

\item{\code{fld_prop}}{The proportion of the field to apply experimental
or optimum check rates to.}

\item{\code{expvar}}{Experimental variable to optimize, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that is experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{conv}}{The conversion factor between lbs of the input to the
units of the as-applied input (i.e. lbs N/ac to unit input/ac)}

\item{\code{base_rate}}{The rate to apply between the experimental rates
and the field edge, or as check rates in the prescription selected
option.}

\item{\code{rx_for_year}}{Provide the year that experiment
is made for. Used for labeling outputs.}

\item{\code{out_path}}{Provide the path to the folder in which to store and
save figures and the prescription Type NA to not create any folders.
You will not be able to save any outputs. (Note, even if a path is provided,
the user can pass FALSE as the sole argument to the 'setupOP' method
to prevent the creation of folders. This will automatically prevent
any plots to be saved.).}

\item{\code{SAVE}}{Logical, whether to save figures and the experiment
Autofilled to FALSE if a user selects NA in the 'out_path' or is NULL.
Autofilled to TRUE otherwise.}

\item{\code{fieldname}}{If the user is creating a new experiment, provide or
select the fieldname of the field to use. The field list is from the
available fields in the database for experimentation.}

\item{\code{farmername}}{If the user is creating a new experiment, provide or
select the name of the farmer that owns or manages the field(s) that
an experiment is going to be generated for.}

\item{\code{exp_rate_length}}{Provide the length of experimental rates to apply.
This applies to new experiments and experimental prescriptions. This
represents the equipment constraints of the farmer. In the case of the
experimental prescription, this number of rates does not include the
number of rates for the optimized base map, so take your selections for
the management scenario and number of optimum rates into account.}

\item{\code{exp_rates}}{Provide a vector of experimental rates equal to the
number of experimental rates provided in 'exp_rate_length'. This is
required for all new experiments, however can be left to null for
experimental prescriptions if experimental rates should be generated
based on gaps in optimum rates.}

\item{\code{exp_rates_prop}}{Provide proportions (0 - 1) for the length
of experimental rates provided in 'exp_rate_length'. This is required
for all new experiments and experimental prescriptions.}

\item{\code{strat_dat}}{If the user is creating a new experiment with stratification
they must apply the data that they are stratifying on. This provides an
identifier for the data being used for the stratification. These must be codes
used in the OFPE workflow sucha as 'yld', 'pro', 'aa_n', 'aa_sr', etc.}

\item{\code{strat_dat_year}}{Follows the same structure of 'strat_dat', however contains
the years to use for stratification of each of the data types for each of the fields.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An initialized ExpGen R6 class object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-executeOutput"></a>}}
\if{latex}{\out{\hypertarget{method-executeOutput}{}}}
\subsection{Method \code{executeOutput()}}{
Method for calling the execution method of the experiment
generator. This randomly applies the experimental rates across
the field. If the user selected stratification data, these are
used for stratification during the random placement.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpGen$executeOutput()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{All parameters supplied upon initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A completed experiment table containing the output.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExpGen$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
