% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFuns.R
\name{findBadParms}
\alias{findBadParms}
\title{Identify parameters inappropriate for use in the OFPE model fitting.}
\usage{
findBadParms(parm_df, dat)
}
\arguments{
\item{parm_df}{Data.frame with a column for the parameters to check, named 'parms',
a column called 'bad_parms' filled with logical responses on whether to omit the
parameter from the model fitting process, and columns labeled 'mean' and 'sd' to
populate in the 'bad_parm' identification process. Other columns may be included
but these columns and specific column names are required.}

\item{dat}{Data.table with the data to check validity of parameters in. Typically
the training data that the model will be fit on.}
}
\value{
Updated 'parm_df' table.
}
\description{
Identifies whether the parameters listed in a data.frame ('parm_df$parms'),
meet the criteria to be omitted from the model, making it a 'bad_parm'. The criteria
for this is over 30\% of data for a given year missing for a parameter or a
standard deviation of zero, indicating singularity. These will lead to errors in the
model fitting process and subsequent predictions. It is a conservative parameter
selection step that promotes the automation of the process.
}
