% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAM.R
\name{GAM}
\alias{GAM}
\title{R6 Class for GAM modeling of crop responses}
\description{
R6 class using Generalized Additive Models (GAM) to fit a crop
response model with the experimental variable and remotely sensed covariate
data. This class is initialized with a named list of training (named 'trn')
and validation (named 'val') datasets, the response variable, the experimental
variable, and the means of the centered data.

The initialization creates a data frame ('parm_df') containing the parameter names,
the k value to use for the GAM, the mean and standard deviation, and whether it meets
the criteria to be omitted from the model, making it a 'bad_parm'. The criteria for
this is over 30\% of data for a given year missing for a parameter or a standard
deviation of zero, indicating singularity.

The process then creates a formula for a final model with all parameters that
are not considered 'bad'. This is used to fit the final model that is returned
to the user for use in the simulation to predict the response under varying
rates of the experimental variable.

The 'saveDiagnostics' method include residuals vs. fitted, normal QQ-
plots, etc. The fitting process also prepares data for validation plots in the
'ModClass' R6 class. This includes predicting observations in the validation
dataset, making a unique id using the year and fieldname, uncentering data, and
identifying a field name to use for plotting that reflects all fields in the
dataset.
}
\seealso{
\code{\link{ModClass}} for the class that calls the ModClass interface,
\code{\link{NonLinear_Logistic}} for the alternative class that fits a
non-linear logistic model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dat}}{Named list of traning (named 'trn') and validation (named 'val')
datasets with the response, experimental, and remotely sensed variables.}

\item{\code{respvar}}{Character, the response variable of interest.}

\item{\code{expvar}}{Character, the experimental variable of interest.}

\item{\code{num_means}}{List for each year in the data with a named vector of the means
for each numerical covariate, including the experimental variable. This is used for
converting centered data back to the original form. The centering process does not
center three numerical variables; the x and y coordinates, and the response variable
(yld/pro). This is for the data specified from the analysis data inputs (grid specific).}

\item{\code{mod}}{Fitted GAM.}

\item{\code{form}}{Final GAM formula.}

\item{\code{parm_df}}{Data frame of parameter names, starting k for the GAM, and a column
named 'bad_parms' to indicate whether to include in the model formula. Also includes
columns for the mean and standard deviation of each parameter.}

\item{\code{fieldname}}{Unique name for the field(s) analyzed. If multiple fields are used
they are separated by an ampersand, otherwise the singular field name is used. This
is used for plottting.}

\item{\code{mod_type}}{Name of the model of this class, used for plotting.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GAM$new()}}
\item \href{#method-fitMod}{\code{GAM$fitMod()}}
\item \href{#method-predResps}{\code{GAM$predResps()}}
\item \href{#method-saveDiagnostics}{\code{GAM$saveDiagnostics()}}
\item \href{#method-clone}{\code{GAM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
The initialization creates a data frame ('parm_df') containing the parameter names,
the k value to use for the GAM, the mean and standard deviation, and whether it meets
the criteria to be omitted from the model, making it a 'bad_parm'. The criteria for
this is over 30\% of data for a given year missing for a parameter or a standard
deviation of zero, indicating singularity.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$new(dat, respvar, expvar, num_means, init_k = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Named list of training (named 'trn') and validation (named 'val')
datasets with the response, experimental, and remotely sensed variables.}

\item{\code{respvar}}{Character, the response variable of interest.}

\item{\code{expvar}}{Character, the experimental variable of interest.}

\item{\code{num_means}}{List for each year in the data with a named vector of the means
for each numerical covariate, including the experimental variable. This is used for
converting centered data back to the original form. The centering process does not
center three numerical variables; the x and y coordinates, and the response variable
(yld/pro). This is for the data specified from the analysis data inputs (grid specific).}

\item{\code{init_k}}{Optional, provide an initial 'k' value to use for the GAM. If no selection
automatically 50. K is the the dimension of the basis used to represent the smooth term.
Multiple k values will be tested, consider this the upper limit and starting place.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A instantiated 'GAM' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fitMod"></a>}}
\if{latex}{\out{\hypertarget{method-fitMod}{}}}
\subsection{Method \code{fitMod()}}{
Method for fitting the GAM to response variables using experimental and
covariate data.

The fitting begins by taking the data frame ('parm_df') containing the parameter names,
the k value to use for the GAM, the mean and standard deviation, and identifying whether
each parameter meets the criteria to be omitted from the model, making it a 'bad_parm'.
The criteria for this is over 30\% of data for a given year missing for a parameter or a
standard deviation of zero, indicating singularity.

This implements another method for identifying 'bad_parms' and fits an initial k value
to pass into the GAM. it sequentially  tests a series of k values and if it does not
converge after setting k = 1, it is considered a bad parm because it does not induce model
convergence.

The process then creates a formula for a final model with all parameters that
are not considered 'bad'. This is used to fit the final model that is returned
to the user for use in the simulation to predict the response under varying
rates of the experimental variable.

Finally, this method prepares the validation data for plotting by using the model to predict
the response for each of the observations in the validation dataset, uncentering data if
necessary, and identifying a unique field name from the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$fitMod()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{Parameters provided upon class instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A fitted GAM.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predResps"></a>}}
\if{latex}{\out{\hypertarget{method-predResps}{}}}
\subsection{Method \code{predResps()}}{
Method for predicting response variables using data and a model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$predResps(dat, mod)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data for predicting response variables for.}

\item{\code{mod}}{The fitted model to use for predicting the response
variable for each observation in 'dat'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of predicted values for each location in 'dat'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-saveDiagnostics"></a>}}
\if{latex}{\out{\hypertarget{method-saveDiagnostics}{}}}
\subsection{Method \code{saveDiagnostics()}}{
Method for saving diagnostic plots of the fitted model. These include residual
vs. fitted values, normal QQ plots, etc.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$saveDiagnostics(out_path, SAVE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the model fitting process}

\item{\code{SAVE}}{Whether to save diagnostic plots.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Diagnostic plots.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
