% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuns.R
\name{getColorRamp}
\alias{getColorRamp}
\title{Create a color ramp from an OFPE variable.}
\usage{
getColorRamp(var)
}
\arguments{
\item{var}{Variable for which to create a color ramp.}
}
\value{
Color ramp appropriate for variable.
}
\description{
Function for returning a color ramp appropriate for
mapping a supplied variable. For example, yield is mapped from
red to green corresponding to low to high.
}
