% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{getExpReps}
\alias{getExpReps}
\title{Apply experimental rates across the field}
\usage{
getExpReps(exp_cells_length, exp_rates_prop)
}
\arguments{
\item{exp_cells_length}{The length of available cells for applying
experimental rates to.}

\item{exp_rates_prop}{Provide proportions (0 - 1) for the length
of experimental rates provided in 'exp_rate_length'. This is required
for all new experiments and experimental prescriptions.}
}
\value{
The number of cells to apply each rate to.
}
\description{
This function gets the number of cells to apply
experimental rates to. Takes the length of available cells for
experimentation and the proportion of the experiment dedicated
to each experimental rate.
}
