% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{getFldBound}
\alias{getFldBound}
\title{Get field(s) boundaries from OFPE database.}
\usage{
getFldBound(fieldname, db, farmername)
}
\arguments{
\item{fieldname}{The field(s) names(s) present in the database to get
field boundaries for.}

\item{db}{Connection to an OFPE formatted database.}

\item{farmername}{Provide the name of the farmer that owns or manages the
field(s) that an prescription is going to be generated for.}
}
\value{
A field boundary sf object.
}
\description{
Gather and combine field boundaries from one or
more selected fields from an OFPE database. Fields must be from
the same farmer.
}
