% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{getRxGrid}
\alias{getRxGrid}
\title{Return a grid for creating experiments and prescriptions.}
\usage{
getRxGrid(
  db,
  rx_dt,
  farmername,
  fieldname,
  trt_length,
  boom_width,
  unq_fieldname
)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{rx_dt}{Data frame that contains the coordinates of locations to apply
experimental inputs. For an experiment, these are the centroids of the grid
made to aggregate data for the field.}

\item{farmername}{If the user is creating a new experiment, provide or
select the name of the farmer that owns or manages the field(s) that
an experiment is going to be generated for.}

\item{fieldname}{If the user is creating a new experiment, provide or
select the fieldname of the field to use. The field list is from the
available fields in the database for experimentation.}

\item{trt_length}{Length, in meters, for which to apply treatments.}

\item{boom_width}{The width of the sprayer boom or spreader.}

\item{unq_fieldname}{Unique fieldname for the field(s) used for the experiment. This
concatenates multiple fields with an ampersand. Used for labelling.}
}
\value{
NULL, grid data stored in database in 'rxgrids' and 'rxgridtemp'.
}
\description{
Returns a grid of the user specified dimensions combined
with data passed in by the user. The data available for use as an
experiment is clipped to the grid, removing areas between the field edge
and the grid within the field. This is used to apply the experimental
or check rates.
}
