% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hush.R
\name{hush}
\alias{hush}
\title{Truly invisible output}
\source{
\url{https://stackoverflow.com/questions/2723034/suppress-output-of-a-function}
}
\usage{
hush(code)
}
\arguments{
\item{code}{Function or code.}
}
\value{
Execution of function or code.
}
\description{
Suppress all output from function return. Useful for st_write. Credit goes
to "Danny", https://stackoverflow.com/users/699385/danny.
}
