% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{importDat}
\alias{importDat}
\title{Add data to an existing table}
\usage{
importDat(db, dat, schema, dtype, upserts)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{dat}{Data to be imported to the database to form the
new table. Must be 'sp' object.}

\item{schema}{Name of the schema with table.}

\item{dtype}{The name of the table to add data to}

\item{upserts}{Columns in the data to check against for
appending data.}
}
\value{
NULL, data added to table in database.
}
\description{
Adds data to an existing database table.
It is assumed that the geometry column in the data is
named 'geometry'.
}
