% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportGEE.R
\name{ImportGEE}
\alias{ImportGEE}
\title{R6 Class for importing Google Earth Engine data to an OFPE database}
\description{
R6 class for for importing data collected from Google Earth
Engine into an OFPE formatted database.
}
\seealso{
\code{\link{DBCon}} for the database connection class, and
\code{\link{ImportOF}} for the associated on-farm data import
class in the data import step.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{dat_path}}{Path to the Google Drive folder that contains the data
for upload into the OFPE database.}

\item{\code{file_names}}{Names of the files in the Google Drive folder.}

\item{\code{overwrite}}{Logical, whether to overwrite the GEE data in the
database.}

\item{\code{farmers}}{The 'farmers' table from the 'all_farms' schema of an
OFPE formatted database.}

\item{\code{farms}}{The 'farms' table from the 'all_farms' schema of an
OFPE formatted database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportGEE-new}{\code{ImportGEE$new()}}
\item \href{#method-ImportGEE-executeUpload}{\code{ImportGEE$executeUpload()}}
\item \href{#method-ImportGEE-.uploadGEE}{\code{ImportGEE$.uploadGEE()}}
\item \href{#method-ImportGEE-.getInfo}{\code{ImportGEE$.getInfo()}}
\item \href{#method-ImportGEE-.addColsToDB}{\code{ImportGEE$.addColsToDB()}}
\item \href{#method-ImportGEE-clone}{\code{ImportGEE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportGEE-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportGEE-new}{}}}
\subsection{Method \code{new()}}{
Create a database importer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportGEE$new(dbCon, dat_path, overwrite)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{dat_path}}{Path to the Google Drive folder that contains the data
for upload into the OFPE database.}

\item{\code{overwrite}}{Logical, whether to overwrite the GEE data in the
database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'ImportGEE' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportGEE-executeUpload"></a>}}
\if{latex}{\out{\hypertarget{method-ImportGEE-executeUpload}{}}}
\subsection{Method \code{executeUpload()}}{
Google Earth Engine data import execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportGEE$executeUpload()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because of class instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
See 'GEE' table in database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportGEE-.uploadGEE"></a>}}
\if{latex}{\out{\hypertarget{method-ImportGEE-.uploadGEE}{}}}
\subsection{Method \code{.uploadGEE()}}{
Uploads data to an OFPE formatted database from a Google Drive folder.
The dot indicates that this function would be private if not for
documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportGEE$.uploadGEE(FILE, overwrite, db)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FILE}}{Name of the data file to upload to the database.}

\item{\code{overwrite}}{Logical, whether to overwrite when data already present.}

\item{\code{db}}{Connection to an OFPE formatted database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
See 'GEE' table in database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportGEE-.getInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ImportGEE-.getInfo}{}}}
\subsection{Method \code{.getInfo()}}{
Function for identifying information from the filenames exported from
Google Drive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportGEE$.getInfo(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the data file to upload to the database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Raster information.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportGEE-.addColsToDB"></a>}}
\if{latex}{\out{\hypertarget{method-ImportGEE-.addColsToDB}{}}}
\subsection{Method \code{.addColsToDB()}}{
Adds columns to the data file importing to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportGEE$.addColsToDB(info, name, db)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{info}}{Object holding all necessary information about the data.}

\item{\code{name}}{The name of the column to add.}

\item{\code{db}}{The connection to the OFPE database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Columns added to database table.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportGEE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportGEE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportGEE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
