% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{importMulti}
\alias{importMulti}
\title{Add data to an existing table}
\usage{
importMulti(db, dat, schema, dtype)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{dat}{Data to be imported to the database to form the
new table. Must be 'sp' object.}

\item{schema}{Name of the schema with target table.}

\item{dtype}{The name of the table to add data to.}
}
\value{
Data uploaded to the database.
}
\description{
Adds data to an existing database table.
It is assumed that the geometry column in the data is
named 'geometry'.
}
