% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{importNewDat}
\alias{importNewDat}
\title{Create a new table in an OFPE database}
\usage{
importNewDat(db, dat, schema, dtype)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{dat}{Data to be imported to the database to form the
new table. Must be 'sp' object.}

\item{schema}{Name of the schema to make.}

\item{dtype}{The name of the table to make.}
}
\value{
NULL, a new table created in the database.
}
\description{
Create a new table in an OFPE database
by inserting a 'sp' data object from R. It is assumed
that the geometry column in the data is named 'geometry'.
}
