% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportOF.R
\name{ImportOF}
\alias{ImportOF}
\title{R6 Class for importing on-farm data to an OFPE database}
\description{
R6 class for for importing data collected on-farms data into an
OFPE formatted database.
}
\seealso{
\code{\link{DBCon}} for the database connection class, and
\code{\link{ImportGEE}} for the associated Google Earth Engine data import
class in the data import step.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{dat_path}}{Path to the folder that contains the data for upload into
the OFPE database.}

\item{\code{file_names}}{Names of the files in the data folder.}

\item{\code{status}}{Data frame with a status update on the completion
of each file's data import.}

\item{\code{farmers}}{The 'farmers' table from the 'all_farms' schema of an
OFPE formatted database.}

\item{\code{farms}}{The 'farms' table from the 'all_farms' schema of an
OFPE formatted database.}

\item{\code{fields}}{The 'fields' table from the 'all_farms' schema of an
OFPE formatted database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImportOF-new}{\code{ImportOF$new()}}
\item \href{#method-ImportOF-executeUpload}{\code{ImportOF$executeUpload()}}
\item \href{#method-ImportOF-.uploadData}{\code{ImportOF$.uploadData()}}
\item \href{#method-ImportOF-.impDat}{\code{ImportOF$.impDat()}}
\item \href{#method-ImportOF-.makeSptl}{\code{ImportOF$.makeSptl()}}
\item \href{#method-ImportOF-.oldDatClean}{\code{ImportOF$.oldDatClean()}}
\item \href{#method-ImportOF-.findInfo}{\code{ImportOF$.findInfo()}}
\item \href{#method-ImportOF-.fixCols}{\code{ImportOF$.fixCols()}}
\item \href{#method-ImportOF-.uploadFun}{\code{ImportOF$.uploadFun()}}
\item \href{#method-ImportOF-clone}{\code{ImportOF$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-new}{}}}
\subsection{Method \code{new()}}{
Create a database importer object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$new(dbCon, dat_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{dat_path}}{Path to the folder that contains the data for upload into
the OFPE database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'ImportOF' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-executeUpload"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-executeUpload}{}}}
\subsection{Method \code{executeUpload()}}{
On-farm data import execution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$executeUpload()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because of class instantiation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
See 'status' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.uploadData"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.uploadData}{}}}
\subsection{Method \code{.uploadData()}}{
Uploads data to an OFPE formatted database following a chain of methods
and returning the status of the upload. The dot indicates that this
function would be private if not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.uploadData(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the data file to upload to the database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
See 'status' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.impDat"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.impDat}{}}}
\subsection{Method \code{.impDat()}}{
Imports data for uploading to an OFPE database. If data is a shapefile
imported directly using 'sf' package, otherwise brings in data as .csv.
The dot indicates that this function would be private if not for
documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.impDat(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the data file to upload to the database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Imported data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.makeSptl"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.makeSptl}{}}}
\subsection{Method \code{.makeSptl()}}{
Makes data that was imported for uploading into spatial data. This only
applies to non-shapefiles, as they are already spatial. This is for the
.csv protein or SSURGO data that is imported. The dot indicates that this
function would be private if not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.makeSptl(FILE, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FILE}}{The imported data from the .impDat method.}

\item{\code{name}}{The filename for the data imported from the .impDat method.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data in spatial format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.oldDatClean"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.oldDatClean}{}}}
\subsection{Method \code{.oldDatClean()}}{
Cleans up any column names that might mess with new ones that are
created in the data workflow. This is mainly to prevent issues when using
data from Phil's database. The dot indicates that this would be private if
not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.oldDatClean(FILE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FILE}}{The imported data for upload.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data without old columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.findInfo"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.findInfo}{}}}
\subsection{Method \code{.findInfo()}}{
Identifies distinguishing features of the data, such as a field name if
present, the corresponding farmer, and the year the data is collected.
The dot indicates that this would be private if not for documentations
sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.findInfo(FILE, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FILE}}{The imported data for upload.}

\item{\code{name}}{The name of the file for upload.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data without old columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.fixCols"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.fixCols}{}}}
\subsection{Method \code{.fixCols()}}{
Goes through each file and replaces any column names with more palatable
formats such as lower case and without any special characters. The dot
indicates that this would be private if not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.fixCols(FILE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FILE}}{The imported data for upload.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data with cleaned up column names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-.uploadFun"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-.uploadFun}{}}}
\subsection{Method \code{.uploadFun()}}{
Function for uploading each file to the appropriate section of the
database. This function identifies the data type and uses the information
identified in the data to organize and upload the data. The dot
indicates that this would be private if not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$.uploadFun(FILE, name, db)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{FILE}}{The imported data for upload.}

\item{\code{name}}{The name of the file for upload.}

\item{\code{db}}{Connection to an OFPE database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data uploaded to the database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImportOF-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImportOF-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImportOF$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
