% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadExtensions.R
\name{loadExtensions}
\alias{loadExtensions}
\title{Load extensions for a new OFPE database}
\source{
\url{https://trac.osgeo.org/postgis/wiki/UsersWikiCreateFishnet}
}
\usage{
loadExtensions(db, postgis_version)
}
\arguments{
\item{db}{Connection to a database.}

\item{postgis_version}{character PostGIS version installed.}
}
\value{
None.
}
\description{
Loads PostGIS tools and a function for generating a net across
an area of interest. From PostGIS website, see source.
}
