% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{makeBaseRate}
\alias{makeBaseRate}
\title{Trim excess experimental rates}
\usage{
makeBaseRate(
  db,
  fieldname,
  unique_fieldname,
  base_rate,
  boom_width,
  trt_length,
  farmername,
  mgmt_scen
)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{fieldname}{If the user is creating a new experiment, provide or
select the fieldname of the field to use. The field list is from the
available fields in the database for experimentation.}

\item{unique_fieldname}{Unique fieldname for the field(s) used for the experiment. This
concatenates multiple fields with an ampersand. Used for labeling.}

\item{base_rate}{The rate to apply between the experimental rates
and the field edge, or as check rates in the prescription selected
option.}

\item{boom_width}{The width of the sprayer boom or spreader.}

\item{trt_length}{Length, in meters, for which to apply treatments.}

\item{farmername}{If the user is creating a new experiment, provide or
select the name of the farmer that owns or manages the field(s) that
an experiment is going to be generated for. Must be same for all
fields.}

\item{mgmt_scen}{If the user is creating a prescription or experimental
prescription, they must provide the management scenario to use for their
prescription. The user can choose from the management options listed in
the SimClass. The options are 'SSOPT': site-specific optimized rates,
'FFOPT': full-field optimum uniform rate, 'FS': farmer selected uniform
rate, 'Min': applying the least intensive input rates (i.e. 0 lbs N/ac, or
25 lbs seed/ac), 'Opp' is omitted because this strategy is the least intensive
input rate in conventional system types, and the farmer selected rate for
organic systems, both of which are already provided. Default to 'base' for new
experiments, override if making a prescription.}
}
\value{
A 'sf' object with field boundaries and the base rate.
}
\description{
This function takes field boundaries from the OFPE
database, sets the base rate, and adds columns required to merge
with the experiment or prescription output.
}
