% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{makeDBcolsText}
\alias{makeDBcolsText}
\title{Set columns in database table to text.}
\usage{
makeDBcolsText(db, dat_names, schema, dtype)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{dat_names}{Names of columns in data that was imported
to the database that are converted to a text data format in the
database.}

\item{schema}{Name of the schema with the table to convert
columns to text in.}

\item{dtype}{The name of the table in the schema
to convert the column names to text for.}
}
\value{
Columns in database table set to text data type.
}
\description{
Set column data type in PostgreSQL database
to text for storage. This is done after data import.
}
