% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{makeRXGrid}
\alias{makeRXGrid}
\title{Create a grid for creating experiments and prescriptions.}
\usage{
makeRXGrid(db, fieldname, trt_length, boom_width, farmername, unq_fieldname)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{fieldname}{If the user is creating a new experiment, provide or
select the fieldname of the field to use. The field list is from the
available fields in the database for experimentation.}

\item{trt_length}{Length, in meters, for which to apply treatments.}

\item{boom_width}{The width of the sprayer boom or spreader.}

\item{farmername}{If the user is creating a new experiment, provide or
select the name of the farmer that owns or manages the field(s) that
an experiment is going to be generated for.}

\item{unq_fieldname}{Unique fieldname for the field(s) used for the experiment. This
concatenates multiple fields with an ampersand. Used for labelling.}
}
\value{
NULL, grid data stored in database in 'rxgrids' and 'rxgridtemp'.
}
\description{
Creates a grid of the user specified dimensions and
stores in the database for later use. The grid is used to place and
apply the experiment or prescription outputs and reflects the dimensions
of the user's equipment and desired treatment lengths.
}
