% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{makeSpatialIndex}
\alias{makeSpatialIndex}
\title{Make spatial index}
\usage{
makeSpatialIndex(db, index_name, schema, dtype)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{index_name}{The name of the spatial index to create.}

\item{schema}{Name of the schema with the table to convert
geometry in.}

\item{dtype}{The name of the table to convert geometry in.}
}
\value{
Spatial index in the OFPE database for the given table.
}
\description{
Create a spatial index on a table in a
PostgreSQL database. Provide the index name which
must be unique across the database, and the schema
and table name.
}
