% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{makeTempBound}
\alias{makeTempBound}
\title{Make a temporary table of field boundaries}
\usage{
makeTempBound(db, fieldname, farmername)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{fieldname}{If the user is creating a new experiment, provide or
select the fieldname of the field to use. The field list is from the
available fields in the database for experimentation.}

\item{farmername}{If the user is creating a new experiment, provide or
select the name of the farmer that owns or manages the field(s) that
an experiment is going to be generated for.}
}
\value{
NULL, temporary field boundary table in 'all_farms'.
}
\description{
Creates a temporary table in the 'all_farms'
schema containing one table with all of the field boundaries
for the fields selected by the user.
}
