% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageDB.R
\name{ManageDB}
\alias{ManageDB}
\title{R6 Class for managing information in the OFPE database}
\description{
R6 class for for adding information to the OFPE database
such as new farmer names, farm names and boundaries, and experimental
field names and boundaries used in OFPE. Users can build and add their
own methods.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{DBCon class object that holds a database connection,
such as PostgreSQLConnection.}

\item{\code{action_list}}{List of lists with the names of the action methods
to execute and their associated parameters.}

\item{\code{do_action}}{List of action methods initialized for execution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ManageDB$new()}}
\item \href{#method-setupActions}{\code{ManageDB$setupActions()}}
\item \href{#method-executeActions}{\code{ManageDB$executeActions()}}
\item \href{#method-clone}{\code{ManageDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a database manager object. The database connection and each
action class list is passed to an internal method that initializes
the specified action classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$new(dbCon, action_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{DBCon class object that holds a database connection,
such as PostgreSQLConnection.}

\item{\code{action_list}}{List of lists with the names of the action methods
to execute and their associated parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'ManageDB' object with initialized action classes.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setupActions"></a>}}
\if{latex}{\out{\hypertarget{method-setupActions}{}}}
\subsection{Method \code{setupActions()}}{
Prepares the R6 action classes supplied by the user in the
'action_list' by running their setup methods that preprocess data
prior to uploading to the OFPE database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$setupActions()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{There are no arguments to this method as the list of
actions was supplied upong initialization of the ManageDB class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Prepared R6 action classes specified by the user.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-executeActions"></a>}}
\if{latex}{\out{\hypertarget{method-executeActions}{}}}
\subsection{Method \code{executeActions()}}{
Executes the upload method of each action class that was passed in by
the user and has been initialzied and set up. These methods upload the
data prepared by each action's setup method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$executeActions()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data uploaded into the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
