% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageFarmers.R
\name{ManageFarmers}
\alias{ManageFarmers}
\title{R6 Class for adding farmer information in OFPE database.}
\description{
R6 class that takes a list with a vector of farmer names and
a database connection object. Adds a unique ID field for each farmer that
is used for internal relationships. If a farmer is already present in the
database, it will be skipped. This follows the ManageDB interface and
includes a setup and execute method that is called from the ManageDB class.
}
\seealso{
\code{\link{DBCon}} for database connection class,
\code{\link{ManageDB}} for the class that calls this class to manage the database.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{farmers}}{Vector of farmer names to add to the database.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ManageFarmers$new()}}
\item \href{#method-setup}{\code{ManageFarmers$setup()}}
\item \href{#method-execute}{\code{ManageFarmers$execute()}}
\item \href{#method-.uploadFarmers}{\code{ManageFarmers$.uploadFarmers()}}
\item \href{#method-clone}{\code{ManageFarmers$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create an ManageFarmers object. The database connection and a vector of
farmer names are used to upload farmer information to the database.
Instantiates the class for running the setup and execute methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageFarmers$new(dbCon, action_item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbCon}}{Database connection object connected to an OFPE formatted
database, see DBCon class.}

\item{\code{action_item}}{List of inputs to the ManageFarmers method, from the
list passed into the ManageDB class. This includes a vector of farmer
names to add to the 'all_farms.farmers' table of an OFPE formatted
database.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new initialized 'ManageFarmers' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
No setup needed for the ManageFarmers class. However is still present to
abide by the ManageDB interface consisting of setup and execute methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageFarmers$setup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-execute"></a>}}
\if{latex}{\out{\hypertarget{method-execute}{}}}
\subsection{Method \code{execute()}}{
Executes the upload of prepared farmer information to the database in the
'all_farms' schema and the 'farmers' table. If a farmer is already
present with the same name and ID, it will be skipped.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageFarmers$execute()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Farmer data uploaded into 'all_farms.farmers' table in the database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-.uploadFarmers"></a>}}
\if{latex}{\out{\hypertarget{method-.uploadFarmers}{}}}
\subsection{Method \code{.uploadFarmers()}}{
Uploads a farmer to the database. Takes the farmer name and adds it to the
'all_farms.farmers' table. Makes sure there is no conflict defined in the
.buildTables method of the BuildDB class. A farmeridx is automatically
generated upon upload by PostgreSQL. The dot indicates that this function
would be private if not for documentations sake.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageFarmers$.uploadFarmers(farmer, db)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{farmer}}{Name of a farmer for upload into 'all_farms.farmers'.}

\item{\code{db}}{Database connection.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Farmer upload into database.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageFarmers$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
