% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{noMatch}
\alias{noMatch}
\title{Get names in source present in target}
\usage{
noMatch(source, target)
}
\arguments{
\item{source}{Character vector to check if present in
target.}

\item{target}{Character vector to check if source elements
are present in.}
}
\value{
Elements of source vector not present in target.
}
\description{
Function for identifying which characters
in a source vector are present in a target vector. Used
to identify which columns don't match in database table
and data column names.
}
