% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noSpecialChar.R
\name{noSpecialChar}
\alias{noSpecialChar}
\title{Remove special characters from strings}
\usage{
noSpecialChar(strng, files)
}
\arguments{
\item{strng}{Character string for cleaning.}

\item{files}{Logical, are the strings filenames? If TRUE, "." not removed,
if FALSE, "." removed.}
}
\value{
Clean character string.
}
\description{
Removes any special characters from a string. Includes all
single special characters as well as ".-" and "__".
}
