% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NRcalcCpp}
\alias{NRcalcCpp}
\title{Function for calculating the net-return for different
management scenarios under the selected economic conditions
for each location in the data frame. For each point the
net-retrun is calculated with the base price, and the
opposite system type's base price. If the rate is equal
to the farmer selected or the minimum experimental rates, the
net-return for those economic conditions are calculated.}
\usage{
NRcalcCpp(
  df,
  Bp,
  B0pd,
  B1pd,
  B2pd,
  CEXP,
  BpOpp,
  FC,
  fs,
  ssAC,
  predInd,
  rr,
  expCol,
  yldCol,
  proCol,
  NRcol,
  NRminCol,
  NRoppCol,
  NRfsCol,
  AAmin
)
}
\arguments{
\item{df}{A matrix with the experimental rate and predicted
responses for each data point in the field.}

\item{Bp}{The base price corresponding to the price for the system
type selected by the user (i.e. conventional or organic).}

\item{B0pd}{The intercept for the protein premium/dockage equation.}

\item{B1pd}{The coefficient for protein in the protein premium/dockage
equation.}

\item{B2pd}{The coefficient for protein squared for the protein
premium/dockage equation.}

\item{CEXP}{The cost of the experimental input.}

\item{BpOpp}{The base price corresponding to the price for the opposite
system type selected by the user (i.e. conventional or organic).}

\item{FC}{The fixed costs associated with production per acre, not including
the experimental input.}

\item{fs}{The farmer selected rate or the rate of the experimental
input that the farmer would have applied uniformly to the field.}

\item{ssAC}{The cost of site-specific technology or variable rate
application.}

\item{predInd}{Indicator for whether protein is included or not
in the net-return calculations. 1 = yes, 0 = no.}

\item{rr}{Number of rows in df.}

\item{expCol}{Indicator for the position of the experimental
variable column in df.}

\item{yldCol}{Indicator for the position of the predicted yield
variable column in df.}

\item{proCol}{Indicator for the position of the predicted protein
variable column in df.}

\item{NRcol}{Indicator for the position of the net-return column
in df.}

\item{NRminCol}{Indicator for the position of the net-return column
in df for the minimum experimental rate applied.}

\item{NRoppCol}{Indicator for the position of the net-return column
in df when net-return calculated with the opposite base price as for the
system.}

\item{NRfsCol}{Indicator for the position of the net-return column
in df for the farmer selected rate.}

\item{AAmin}{The minimum experimental rate in the range of rates to
simulate in (i.e. 0 lbs N/ac).}
}
\value{
df Matrix filled in with net-returns.
}
\description{
Function for calculating the net-return for different
management scenarios under the selected economic conditions
for each location in the data frame. For each point the
net-retrun is calculated with the base price, and the
opposite system type's base price. If the rate is equal
to the farmer selected or the minimum experimental rates, the
net-return for those economic conditions are calculated.
}
