% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUE.R
\name{NUE}
\alias{NUE}
\title{R6 Class for working with nitrogen use efficiency}
\description{
R6 class for fitting the NUE model based on a mass balance
approach using intensive soil and plant biomass sampling outlined in
Hegedus & Ewing, 2022. NUE is calculated as the ratio of total N taken up
by a crop over the amount of N available from fertilization, atmospheric
deposition (negligible), and mineralized N. Mineralized N is estimated from
a general model in Vigil et al. 2002.

The NUE model is fit using open-source covariates aggregated per the
OFPE framework and coinciding with the data used for crop responses models
and in data in the 'sat' tables of farmer's aggregated schemas. The NUE
model is used when the optimization method in 'SimClass' is set to 'ecol'.

Upon initialization of this class, the training data with NUE observations are
used to train a support vector regression model characterizing NUE. There are
methods for feeding in data for which to predict NUE on. Additionally, there
is a method for plotting NUE across a covariate, such as N fertilizer.

When used in the 'SimClass', the NUE is calculated at each point in the simulation
dataset for every N rate. When net-return is calculated, a method in this class
can be called to calculate the cost of inefficiency (AN x CN) - (AN x NUE x CN)
which is subtracted from net-return. Then the profit maximizing N rate is found
because it is the rate that maximizes profit and minimizes the cost of N inefficiency.
}
\seealso{
\code{\link{SimClass}} for use of the NUE class in the simulation and
derivation of optimized rates.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{trn_dat}}{Data used for training, built into package.}

\item{\code{mod_type}}{Type of model to use for NUE.}

\item{\code{m}}{Trained NUE model used to make predictions of NUE in new
datasets.}

\item{\code{utm_epsg}}{The code for the UTM zone that the field falls within (e.g. 32612).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{NUE$new()}}
\item \href{#method-nuePred}{\code{NUE$nuePred()}}
\item \href{#method-clone}{\code{NUE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NUE$new(mod_type = "SVR", utm_epsg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mod_type}}{Type of model to use for NUE. Defaults to support vector
regresssion "SVR" but user has options for simple linear regression "SLR",
multiple linear regression "MLR", polynomial regression with a cubic
term on N "PLR", non-linear exponential decay "NLR", generalized additive
modeling "GAM", or random forest regression "RF".}

\item{\code{utm_epsg}}{The code for the UTM zone that the field falls within (e.g. 32612).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A NUE model is fit and initialized in the class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-nuePred"></a>}}
\if{latex}{\out{\hypertarget{method-nuePred}{}}}
\subsection{Method \code{nuePred()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NUE$nuePred(new_dat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_dat}}{New data to make predictions of NUE in.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Predictions of NUE for each observation in the new data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NUE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
