% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsOP.R
\name{ObsOP}
\alias{ObsOP}
\title{R6 Class for plotting outputs of the OFPE simulation.}
\description{
R6 class for generating maps and figures from observed data
in an OFPE database.

This class ...

Uses aggregated not raw data in the database.
can be used to pull data from database

is initialized for field(s) and year(s) observation.
If multiple years the mean for each location in the dataset is used.

can make maps of any variable in aggregated data (uses mean if multiple years)
can crate scatterplots between two variables in data
can create histograms/pirate plots?/boxplots of any variable in data

correlation stuff between yield and protein? map with correlation and overall
r & cv for each?
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{ObsOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
