% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFPE.R
\docType{package}
\name{OFPE}
\alias{OFPE}
\title{OFPE: On-Farm Precision Experiments data management and analysis tools}
\description{
The OFPE package contains tools for supporting the OFPE data framework
developed at Montana State University. This package contains S3 and R6
objects for executing data management and analysis tasks, including;
creating an OFPE formatted spatial database, importing on-farm and
remotely sensed data, aggregating data from disparate sources, data
analysis and optimization of inputs, simulationg of management outcomes,
and generation of prescriptive application maps for optimized inputs.
}
\details{
This package relies on a PostgreSQL database named 'OFPE'. The user is
assumed to have downloaded PostgreSQL and PostGIS. This process is
outlined on the OFPE Website \url{https://paulhegedus.github.io/OFPE-Website/}.

The package follows Google's R Style Guide which follows the Tidyverse Style Guide.
However, there are a couple distinctions from Google's style guide. Rather
than naming functions with 'BigCamelCase', they are named with 'smallCamelCase',
while R6 classes are named with 'BigCamelCase' and assigned to 'smallCamelCase'
objects. Anything besides a class object or function is in 'snake_case.'
}
\section{General}{

General R6 Classes that spread over more than one step.
DBCon, DatClass, ModClass, SimClass

General Function Files.
removeTempFuns, analysisFuns, findUTMzone, legendMap, importFuns,
noSpecialChar, plotFuns, rxFuns
}

\section{Database Creation & Managements}{

See Vignette 1 for database management and updating examples.
The classes and functions used in this step of the OFPE cycle
are listed below.

OFPE database build and management R6 classes.
BuildDB, ManageDB, ManageFarmers, ManageFarms, ManageFields
}

\section{Data Import}{

See the Vignettes 2 for examples importing data from farms and
Google Earth Engine. The classes and functions used in this step
of the OFPE cycle are listed below.

The OFPE data import R6 classes.
ImportOF, ImportGEE

The OFPE data import functions.
setNAtoNaN, makeDBcolsText, convPolyToMulti, makeSpatialIndex,
importNewDat, standardizeColNames, noMatch, tabExist, importDat,
importMulti
}

\section{Data Aggregation}{

See Vignette 3 for examples aggregating data from disparate sources.
The classes and functions used in this step of the OFPE cycle
are listed below.
The OFPE data aggregation R6 classes.
AggDat, AggGEE, AggInputs
}

\section{Data Analysis and Simulation}{

See Vignette 4 for an example on analyzing collected data and simulating
management outcomes. The classes and functions used in this step of the
OFPE cycle are listed below.

The OFPE data analysis and simulation R6 classes.
DatClass, EconDat, ModClass, SimClass, SimOP, RcppExports

The OFPE data analysis and simulation R6 functions.
findBadParms, valPrep, uniqueFieldname, plotMaps, getColorRamp
}

\section{Prescription Generation}{

See Vignette 5 for an example on creating an optimized site-specific variable
rate input prescription. The classes and functions used in this step of the
OFPE cycle are listed below.

The OFPE experiment and prescription generation R6 classes.
RxClass, SimClass, ExpGen, RxGen

The OFPE experiment and prescription generation R6 functions.
getRxGrid, makeTempBound, makeRXGrid, getFldBound, applyExpRates,
getExpReps, trimExpReps, makeBaseRate, trimGrid, makeRx,
uniqueFieldname, plotMaps, getColorRamp
}

\seealso{
\url{https://sites.google.com/site/ofpeframework/}

\url{https://paulhegedus.github.io/OFPE-Website/}

\url{https://github.com/paulhegedus/OFPEDATA.git}

\url{https://google.github.io/styleguide/Rguide.html}

\url{https://style.tidyverse.org}
}
