% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFuns.R
\name{removeNAfromCovars}
\alias{removeNAfromCovars}
\title{Remove NA values from covariate columns.}
\usage{
removeNAfromCovars(dat, covars)
}
\arguments{
\item{dat}{Data.table or data.frame to remove missing values from.}

\item{respvar}{Character vector, the column names to look and remove missing observations
from.}
}
\value{
Data.table with missing observations removed from specified columns.
}
\description{
Function for removing missing values from columns of a data.frame
or data.table. Must pass in the data to remove missing values from and a character
vector of covariate names that correspond to column names in the data. Only observations
from these columns are removed.
}
