% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTempFuns.R
\name{removeTempTables}
\alias{removeTempTables}
\title{Remove temporary tables from 'all_farms' schema}
\usage{
removeTempTables(db)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}
}
\value{
NULL, tables removed in database.
}
\description{
Removes any temporary tables from the 'all_farms' schema of an
OFPE formatted database.
}
