% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendMap.R
\name{scale_bar}
\alias{scale_bar}
\title{Enables drawing of a scale bar on a ggplot object, and orientation arrow (optional)}
\source{
https://github.com/3wen/legendMap.git
}
\usage{
scale_bar(
  lon,
  lat,
  distance_lon,
  distance_lat,
  distance_legend,
  dist_unit = "km",
  rec_fill = "white",
  rec_colour = "black",
  rec2_fill = "black",
  rec2_colour = "black",
  legend_colour = "black",
  legend_size = 3,
  orientation = TRUE,
  arrow_length = 500,
  arrow_distance = 300,
  arrow_north_size = 6
)
}
\arguments{
\item{lon}{Longitude of the bottom left point of the first rectangle to draw.}

\item{lat}{Latitude of the bottom left point of the first rectangle to draw.}

\item{distance_lon}{Length of each rectangle.}

\item{distance_lat}{Width of each rectangle.}

\item{distance_legend}{Distance between rectangles and legend texts.}

\item{dist_unit}{Units of distance "km" (kilometers) (default), "nm"
(nautical miles), "mi" (statute miles).}

\item{rec_fill}{Filling colour of the first rectangle (default to white).}

\item{rec_colour}{Colour of the first rectangle (default to black).}

\item{rec2_fill}{Filling colour of the second rectangle (default to black).}

\item{rec2_colour}{Colour of the second rectangle (default to black).}

\item{legend_colour}{Legend colour (default to black).}

\item{legend_size}{Legend size (default to 3).}

\item{orientation}{Logical, if TRUE (default), adds an orientation arrow to
the plot.}

\item{arrow_length}{Length of the arrow (default to 500 km).}

\item{arrow_distance}{Distance between the scale bar and the bottom of
the arrow (default to 300 km)}

\item{arrow_north_size}{Size of the "N" letter (default to 6).}
}
\value{
Enables drawing of a scale bar on a ggplot object,
and optionally an orientation arrow. Use in ggplot plot code.
}
\description{
Function from 3wen/legendMap repo for 'adding a North arrow and/or a
scale bar to maps done using ggplot2'. All credit for this function goes to
Ewen Gallic. All notation and documentation is from Ewen Gallic.

This function Enables drawing of a scale bar on a ggplot object, and orientation
arrow (optional). See https://github.com/3wen/legendMap.git for examples. Utilizes
create_scale_bar(), and create_orientation_arrow().
}
