% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{selectDatCpp}
\alias{selectDatCpp}
\title{Function for subsetting remotely sensed data from the user
selected source. This function takes a matrix with all the
remotely sensed data and returns a matrix with the selected
data. This function exists because data such as precipitation and
vegetation indices are collected from multiple sources, and
the user can select the source based on their data preferences.}
\usage{
selectDatCpp(
  dat,
  rr,
  prec_cy_1,
  prec_cy_2,
  prec_py_1,
  prec_py_2,
  gdd_cy_1,
  gdd_cy_2,
  gdd_py_1,
  gdd_py_2,
  veg_cy_1,
  veg_cy_2,
  veg_py_1,
  veg_py_2,
  veg_2py_1,
  veg_2py_2
)
}
\arguments{
\item{dat}{A matrix with all of the remotely sensed data to
select from.}

\item{rr}{The number of rows in dat and df (must be the same).}

\item{prec_cy_1}{The column index of the preferred precipitation source for
the current year.}

\item{prec_cy_2}{The column index of the secondary precipitation source for
the current year.}

\item{prec_py_1}{The column index of the preferred precipitation source for
the previous year.}

\item{prec_py_2}{The column index of the secondary precipitation source for
the previous year.}

\item{gdd_cy_1}{The column index of the preferred growing degree day source
for the current year.}

\item{gdd_cy_2}{The column index of the secondary growing degree day source
for the current year.}

\item{gdd_py_1}{The column index of the preferred growing degree day source
for the previous year.}

\item{gdd_py_2}{The column index of the secondary growing degree day source
for the previous year.}

\item{veg_cy_1}{The column index of the preferred vegetation index source
for the current year.}

\item{veg_cy_2}{The column index of the secondary vegetation index source
for the current year.}

\item{veg_py_1}{The column index of the preferred vegetation index source
for the previous year.}

\item{veg_py_2}{The column index of the secondary vegetation index source
for the previous year.}

\item{veg_2py_1}{The column index of the preferred vegetation index source
for two years prior.}

\item{veg_2py_2}{The column index of the secondary vegetation index source
for two years prior.}
}
\value{
df Returns a completed table with selected data.
}
\description{
The user also must pass in the column index of the data passed
in that corresponds to the primary and secondary source for each
variable.
}
