% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{setNAtoNaN}
\alias{setNAtoNaN}
\title{Set NA values to NaN}
\usage{
setNAtoNaN(dat)
}
\arguments{
\item{dat}{Data to check for NA's and convert to NaN's.}
}
\value{
Data with NA values converted to NaN.
}
\description{
Set Na values to NaN in all columns of a data.
This converts NA values in R to NaN's compatible with the
PostgreSQL database. This is done prior to import into the
database.
}
