% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInputs.R
\name{SimInputs}
\alias{SimInputs}
\title{R6 Class for storing inputs for analysis, simulations, and prescriptions}
\description{
R6 class for for storing information needed mainly for the
simulation performed post-analysis and pre-prescription building. This
class stores inputs such as the model used for analysis and the subsequent
simulation, the number of iterations for the simulation, and the weather
scenario to simulate. This class also is used for storing inputs related
to the model and simulation outputters.

Inputs can be supplied directly to this class during instantiation, however
this is NOT recommended except for advanced users. It is recommended that the
user supplies the database connection and uses the interactive selection
methods to select user inputs.

This class is used for instantiating the selected model class and associated
outputter, activating and executing the 'LikeYear' class to identify the year
from the past that is most likely to reflect the upcoming year. This class
also is used to instantiate the simulation class.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{SimInputs$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimInputs$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
