% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{standardizeColNames}
\alias{standardizeColNames}
\title{Standardize columns in data and database table}
\usage{
standardizeColNames(db, dat, schema, dtype)
}
\arguments{
\item{db}{Database connection to OFPE formatted database.}

\item{dat}{Data to be imported to the database.}

\item{schema}{Name of the schema where the target table is.}

\item{dtype}{The name of the target table.}
}
\value{
Data with new columns, new columns in database table.
}
\description{
Add columns that are present in the data and not
in the database table to the database table and add the
columns present in the database table not present in the data
to the data.
}
