% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFuns.R
\name{tabExist}
\alias{tabExist}
\title{Check if table exists}
\usage{
tabExist(db, schema, dtype)
}
\arguments{
\item{db}{Connection to OFPE database.}

\item{schema}{Name of the schema where the target table is.}

\item{dtype}{The name of the target table.}
}
\value{
Logical, if table exists in database.
}
\description{
Check if a table exists in the OFPE database
by checking for the user provided schema and table.
}
