% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFuns.R
\name{takeSubset}
\alias{takeSubset}
\title{Take random subset of n samples.}
\usage{
takeSubset(x, respvar, sub_n = 1250)
}
\arguments{
\item{x}{data.frame or data.table.}

\item{respvar}{Character, the response variable column name.}

\item{sub_n}{Numeric, the number of random observations to take.}
}
\value{
A random subset of the data.
}
\description{
Function for taking a random subset of the data. Used for
top-down selection in model fitting to speed up processes. This creates
a representative sample across fields and years if they're present.
}
