% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{trimExpReps}
\alias{trimExpReps}
\title{Trim excess experimental rates}
\usage{
trimExpReps(cell_diff, exp_reps, midpoint)
}
\arguments{
\item{cell_diff}{Vector with the number of cells difference between
the number of rates and the number of available cells for experimentation}

\item{exp_reps}{The number or cells to apply to each
epxerimental rate.}

\item{midpoint}{The median rate number for removing reps from middle
rates before the extremes.}
}
\value{
The number of cells to apply each rate to.
}
\description{
This function trims the number of cells to apply
experimental rates to the number of available experimental rates.
Due to rounding errors with proportions, more cells are selected
than available. This removes reps from median rates first before
moving to the extremes because it is assumed that the majority of
optimum or other base rates will be around the median experimental
rate. Additionally, when the applicator is moving from low to high
or high to low rates, these middle rates will be applied. The minimum
number of cells to apply a rate to is 1.
}
