% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFuns.R
\name{valPrep}
\alias{valPrep}
\title{Prepare validation data for plotting.}
\usage{
valPrep(dat, respvar, expvar, num_means)
}
\arguments{
\item{dat}{Data.table with the data to check validity of parameters in. Typically
the validation data that the model will be tested on.}

\item{respvar}{Character, the response variable of interest.}

\item{expvar}{Character, the experimental variable of interest.}

\item{num_means}{Named vector of the means for each numerical covariate, including
the experimental variable. This is used for converting centered data back to the
original form. The centering process does not center three numerical variables; the
x and y coordinates, and the response variable (yld/pro). This is for the data specified
from the analysis data inputs (grid specific).}
}
\value{
Updated 'parm_df' table.
}
\description{
Adds a column to the data.table/data.frame with a unique year and field
descriptor by concatenating year and field, separated by a '.', corrects the experimental
variable values back to the observed by reversing the centering. If the data was not centered
there is no change. Removes missing values from the response variable and predicted response
variable columns.
}
