% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findUTMzone.R
\name{findUTMzone}
\alias{findUTMzone}
\title{Identify the Universal Transverse Mercator coordinate system zone}
\usage{
findUTMzone(db, farm = NULL, fieldname = NULL)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{farm}{Either a farm id (integer), or a farm name.}

\item{fieldname}{Name of the field for aggregation.}
}
\value{
WGS84 UTM code (i.e. 32612, 32614).
}
\description{
Function for querying the database to identify the UTM zone
from the corresponding farm for the field the user is working with. This
function is used many times and requires a database connection and either
a farm name or farm id, or a field name. The returned value
is the UTM EPSG code for the WGS84 datum.
}
