% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{makeRXGrid}
\alias{makeRXGrid}
\title{Create a grid for creating experiments and prescriptions.}
\usage{
makeRXGrid(
  db,
  fieldname,
  trt_length,
  boom_width,
  farmername,
  unique_fieldname,
  mgmt_scen
)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{fieldname}{Provide the name of the field for the experiment or prescription.}

\item{trt_length}{Length, in meters, for which to apply treatments.}

\item{boom_width}{The width of the sprayer boom or spreader.}

\item{farmername}{The name of the farmer that manages the field.}

\item{unique_fieldname}{Unique fieldname for the field(s) used for the experiment. This
concatenates multiple fields with an ampersand. Used for labeling.}

\item{mgmt_scen}{If the user is creating a prescription or experimental
prescription, they must provide the management scenario to use for their
prescription. The user can choose from the management options listed in
the SimClass. The options are 'SSOPT': site-specific optimized rates,
'FFOPT': full-field optimum uniform rate, 'FS': farmer selected uniform
rate, 'Min': applying the least intensive input rates (i.e. 0 lbs N/ac, or
25 lbs seed/ac), 'Opp' is omitted because this strategy is the least intensive
input rate in conventional system types, and the farmer selected rate for
organic systems, both of which are already provided.}
}
\value{
NULL, grid data stored in database in 'rxgrids' and 'rxgridtemp'.
}
\description{
Creates a grid of the user specified dimensions and
stores in the database for later use. The grid is used to place and
apply the experiment or prescription outputs and reflects the dimensions
of the user's equipment and desired treatment lengths. The 'boom_width',
or width of the applicator, is multiplied by two, to make treatments with
two passes of the equipment.
}
