% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{makeTempBound}
\alias{makeTempBound}
\title{Make a temporary table of field boundaries}
\usage{
makeTempBound(db, fieldname, farmername)
}
\arguments{
\item{db}{Connection to an OFPE formatted database.}

\item{fieldname}{Provide the field name for the experiment or prescription.}

\item{farmername}{Provide the name of the farmer that manages the field.}
}
\value{
NULL, temporary field boundary table in 'all_farms'.
}
\description{
Creates a temporary table in the 'all_farms'
schema containing one table with all of the field boundaries
for the fields selected by the user.
}
