% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ManageDB.R
\name{ManageDB}
\alias{ManageDB}
\title{R6 Class for managing information in the OFPE database}
\source{
\url{https://trac.osgeo.org/postgis/wiki/UsersWikiCreateFishnet}
}
\description{
R6 class for for adding information to the OFPE database
such as new farmer names, farm names and boundaries, and experimental
field names and boundaries used in OFPE. Users can build and add their
own methods.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db}}{Database connection, such as PostgreSQLConnection.}

\item{\code{action_list}}{List of lists with the names of the action methods
to execute and their associated parameters.}

\item{\code{do_action}}{List of action methods initialized for execution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ManageDB$new()}}
\item \href{#method-executeAction}{\code{ManageDB$executeAction()}}
\item \href{#method-initializeActions}{\code{ManageDB$initializeActions()}}
\item \href{#method-clone}{\code{ManageDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a database manager object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$new(db = NA, action_list = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Database connection.}

\item{\code{action_list}}{List of lists with the names of the action methods
to execute and their associated parameters.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'ManageDB' object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-executeAction"></a>}}
\if{latex}{\out{\hypertarget{method-executeAction}{}}}
\subsection{Method \code{executeAction()}}{
Execute database manager action methods that the user specifies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$executeAction()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{None}}{No arguments needed because they are provided during class
initialization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initializeActions"></a>}}
\if{latex}{\out{\hypertarget{method-initializeActions}{}}}
\subsection{Method \code{initializeActions()}}{
Initializes the a
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$initializeActions()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Connection to a database.}

\item{\code{postgis_version}}{PostGIS version installed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Enabled database extensions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ManageDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
