% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsOP.R
\name{ObsOP}
\alias{ObsOP}
\title{R6 Class for plotting outputs of the OFPE simulation.}
\description{
R6 class for generating maps and figures from observed data
in an OFPE database. This class can be initialized with arguments required
to run the methods for acquiring or making figures, or can be initialized solely
with a 'create' argument, described below. When an empty 'ObsOP' class is initialized,
the methods must have all arguments passed, whereas if the user initializes the
'ObsOP' class with arguments, methods can be executed with minimal other arguments.

It is recommended to initialize one 'ObsOP' class for each dataset desired for
analysis. Multiple fields will be treated as one, and only one year is allowed
for gathering.

This class can be used to import data from an OFPE database for data
exploration, quality control, and general use. The user can fetch data
by providing specific arguments, or by providing a SQL statement for
customized fetching from the database. Data can be returned to an object
or can be used internally with the methods below.

Methods are available for combining data from datasets by kriging. This can be
useful for putting protein and yield in the same data frame or either/both on
data from a 'sat' table on grid points. This would also be required to use the
method for calculating net-return with yield and protein.

This class also contains methods for mapping and plotting variables from
user specified data. The user can map 10m rasters or the points of the
specified data, can create scatterplots of two variables, or can create
histograms of a variable. These figures are returned from the methods but also can be saved to a user
specified locations. This is an R6 class so all parameters needed for
generating figures can be supplied upon class initialization. However, to
use more generally, the class can be initialized and figure making methods
can be run individually.
\subsection{TODO: correlation/CV maps, summary statistics}{
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{out_path}}{Optional, path to folder to save outputs. Also serveds
as a logical for whether to create the 'Output' folder and save
figures, maps, and tables. If this argument is left null, the user will
have to provide the path to save outputs for each individual method.}

\item{\code{SAVE}}{Whether to save outputs. This is set based off of the user
providing a path to a folder to make outputs.}

\item{\code{unique_fieldname}}{Unique name for the field(s) used in the simulation.
This is used for saving and labeling of outputs. e.g. "sec35middle" or
"sec1east & sec1west". This must match the 'unique_fieldname' used to
save the simulation outputs.}

\item{\code{utm_fieldname}}{The name of the field (corresponding to the name in the
OFPE database) that should be used for identifying the UTM zone. In the
case of multiple fields (i.e. "sec1east & sec1west") choose one because
their UTM code will be the same.}

\item{\code{db}}{Connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{fieldname}}{The name of the field(s) to gather data for. Must
match the format in the OFPE database.}

\item{\code{year}}{The year to gather data for.}

\item{\code{dat_tab}}{The table that the data is in within the aggregated schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'pro', or 'sat').}

\item{\code{GRID}}{Select whether to gather gridded or observed data
locations. See the 'AggInputs' class for more information on the 'GRID'
option. Provide either "grid" or "obs".}

\item{\code{dat_used}}{Option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection.}

\item{\code{orig_file}}{The original file name of the data to gather.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ObsOP$new()}}
\item \href{#method-fetchAggDat}{\code{ObsOP$fetchAggDat()}}
\item \href{#method-fetchRawDat}{\code{ObsOP$fetchRawDat()}}
\item \href{#method-krigeDat}{\code{ObsOP$krigeDat()}}
\item \href{#method-calcNR}{\code{ObsOP$calcNR()}}
\item \href{#method-plotObsMaps}{\code{ObsOP$plotObsMaps()}}
\item \href{#method-plotScatters}{\code{ObsOP$plotScatters()}}
\item \href{#method-plotHist}{\code{ObsOP$plotHist()}}
\item \href{#method-clone}{\code{ObsOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Use this class to import data from an OFPE database, save figures and
explore observed data. This class can be initialized with a arguments required
for executing the methods of the class or with only a 'create' parameter,
see below. If the user supplies all arguments to the class, they can
typically use the methods with one or two arguments. A database connection
must be supplied to import data from the database. If the user imports data,
it is initialized into the class.

The logical argument, 'create', gives the user a choice to select whether to save
figures and maps. Depending on this choice, the output folders will be generated.
This is done by a private method that sets up the output location for
the figures that the model produces. This will not overwrite any previously
generated diagnostic or validation plots from the ModClass. Functions for
plots and maps all have save options which must be accompanied by a folder
path. The folder created is named 'Outputs'. This folder contains a
folder called 'ObsMaps' which contains maps of observed data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$new(
  out_path = NULL,
  farmername = NULL,
  fieldname = NULL,
  year = NULL,
  dat_tab = NULL,
  GRID = NULL,
  db = NULL,
  dat_used = NULL,
  utm_fieldname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_path}}{Optional, path to folder to save outputs. Also serveds
as a logical for whether to create the 'Output' folder and save
figures, maps, and tables. If this argument is left null, the user will
have to provide the path to save outputs for each individual method.}

\item{\code{farmername}}{Optional, the name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{fieldname}}{Optional, the name of the field(s) to gather data for. Must
match the format in the OFPE database.}

\item{\code{year}}{Optional, the year to gather data for.}

\item{\code{dat_tab}}{Optional, the table that the data is in within the aggregated schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'pro', or 'sat').}

\item{\code{GRID}}{Optional, select whether to gather gridded or observed data
locations. See the 'AggInputs' class for more information on the 'GRID'
option. Provide either "grid" or "obs".}

\item{\code{db}}{Optional, connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{dat_used}}{Optional, option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection. Provide either
"decision_point" or "full_year".}

\item{\code{utm_fieldname}}{The name of the field (corresponding to the name in the
OFPE database) that should be used for identifying the UTM zone. In the
case of multiple fields (i.e. "sec1east & sec1west") choose one because
their UTM code will be the same.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Initialized R6 class and a folder created in the path for
model output figures if specified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetchAggDat"></a>}}
\if{latex}{\out{\hypertarget{method-fetchAggDat}{}}}
\subsection{Method \code{fetchAggDat()}}{
Gather data from an aggregated table in an
OFPE database. Provide the farmername, fieldname(s), year(s),
and type of data to gather.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$fetchAggDat(
  db = self$db,
  farmername = self$farmername,
  fieldname = self$fieldname,
  year = self$year,
  dat_tab = self$dat_tab,
  GRID = self$GRID,
  dat_used = self$dat_used
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{fieldname}}{The name of the field(s) to gather data for. Must
match the format in the OFPE database.}

\item{\code{year}}{The year(s) to gather data for.}

\item{\code{dat_tab}}{The table that the data is in within the aggregated schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'pro', or 'sat').}

\item{\code{GRID}}{Select whether to gather gridded or observed data
locations. See the 'AggInputs' class for more information on the 'GRID'
option. Provide either "grid" or "obs".}

\item{\code{dat_used}}{Optional, option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection. Provide either
"decision_point" or "full_year".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with specified data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fetchRawDat"></a>}}
\if{latex}{\out{\hypertarget{method-fetchRawDat}{}}}
\subsection{Method \code{fetchRawDat()}}{
Gather data from a raw table in an
OFPE database. Provide the farmername, the type of data, and the
name of the original file for the data ('orig_file').
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$fetchRawDat(
  db = self$db,
  farmername = self$farmername,
  dat_tab = self$dat_tab,
  orig_file = self$orig_file
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{dat_tab}}{The table that the data is in within the raw schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'aa_n_poly', 'aa_sr', etc.).}

\item{\code{orig_file}}{The original file name of the data to gather.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with specified data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-krigeDat"></a>}}
\if{latex}{\out{\hypertarget{method-krigeDat}{}}}
\subsection{Method \code{krigeDat()}}{
Interpolate data using kriging.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$krigeDat(source_dat, target_dat, var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_dat}}{The data to interpolate from.}

\item{\code{target_dat}}{The data to krige data to.}

\item{\code{var}}{The variable to interpolate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with interpolated data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calcNR"></a>}}
\if{latex}{\out{\hypertarget{method-calcNR}{}}}
\subsection{Method \code{calcNR()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$calcNR()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotObsMaps"></a>}}
\if{latex}{\out{\hypertarget{method-plotObsMaps}{}}}
\subsection{Method \code{plotObsMaps()}}{
This method is for plotting maps of observed variables. The user
must supply the data and variable to map. Other arguments relate to
labeling the map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotObsMaps(
  dat,
  var,
  var_col_name,
  var_label,
  var_main_label = NULL,
  fieldname = self$unique_fieldname,
  year = self$year,
  SAVE = self$SAVE,
  farmername = self$farmername,
  out_path = self$out_path,
  db = self$db,
  utm_fieldname = self$utm_fieldname
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with variables to map. Must include an x and y
column.}

\item{\code{var}}{The label of the variable to map. Used in figure name.}

\item{\code{var_col_name}}{The name of the column of the variable in the
supplied data ('dat').}

\item{\code{var_label}}{The label to be applied to the legend of the map
corresponding to the variable mapped.}

\item{\code{var_main_label}}{The main label to apply to the map.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{year}}{Year of the observed data.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{farmername}}{The name of the farmer that manages the field.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{db}}{Connection to the OFPE database to identify UTM zone.}

\item{\code{utm_fieldname}}{Name of the field for identifying the UTM zone.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggmap' object and maps saved in 'Outputs/Maps' folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotScatters"></a>}}
\if{latex}{\out{\hypertarget{method-plotScatters}{}}}
\subsection{Method \code{plotScatters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotScatters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotHist"></a>}}
\if{latex}{\out{\hypertarget{method-plotHist}{}}}
\subsection{Method \code{plotHist()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotHist()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
