% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsOP.R
\name{ObsOP}
\alias{ObsOP}
\title{R6 Class for plotting outputs of the OFPE simulation.}
\description{
R6 class for generating maps and figures from observed data
in an OFPE database. This class can be initialized with arguments required
to run the methods for acquiring or making figures, or can be initialized empty.
When an empty 'ObsOP' class is initialized, the methods must have all arguments
passed each time, whereas if the user initializes the 'ObsOP' class with arguments
specifying the data selections, the methods can be executed with minimal other arguments.
This is meant to simplify use and allow the user to have multiple ObsOP classes for
different datasets.

It is recommended to initialize one 'ObsOP' class for each dataset desired for
analysis by passing in the specifications for data in the instantiation of the class.
Multiple fields will be treated as one, and only one year is allowed
for gathering.

This class can be used to import data from an OFPE database for data
exploration, quality control, and general use. The user can fetch data
by providing specific arguments, or by providing a SQL statement for
customized fetching from the database. Data can be returned to an object
or can be used internally with the methods below.

Methods are available for combining data from datasets by kriging. This can be
useful for putting protein and yield in the same data frame or either/both on
data from a 'sat' table on grid points. This would also be required to use the
method for calculating net-return with yield and protein.

This class also contains methods for mapping and plotting variables from
user specified data. The user can map 10m rasters or the points of the
specified data, can create scatterplots of two variables, or can create
histograms of a variable. These figures are returned from the methods but also can be saved to a user
specified locations. This is an R6 class so all parameters needed for
generating figures can be supplied upon class initialization. However, to
use more generally, the class can be initialized and figure making methods
can be run individually. TODO: correlation/CV maps, summary statistics
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{out_path}}{Optional, path to folder to save outputs. Also serveds
as a logical for whether to create the 'Output' folder and save
figures, maps, and tables. If this argument is left null, the user will
have to provide the path to save outputs for each individual method.}

\item{\code{SAVE}}{Whether to save outputs. This is set based off of the user
providing a path to a folder to make outputs.}

\item{\code{unique_fieldname}}{Unique name for the field(s) used in the simulation.
This is used for saving and labeling of outputs. e.g. "sec35middle" or
"sec1east & sec1west". This must match the 'unique_fieldname' used to
save the simulation outputs.}

\item{\code{utm_fieldname}}{The name of the field (corresponding to the name in the
OFPE database) that should be used for identifying the UTM zone. In the
case of multiple fields (i.e. "sec1east & sec1west") choose one because
their UTM code will be the same.}

\item{\code{db}}{Connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{fieldname}}{The name of the field(s) to gather data for. Must
match the format in the OFPE database.}

\item{\code{year}}{The year to gather data for.}

\item{\code{dat_tab}}{The table that the data is in within the aggregated schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'pro', or 'sat').}

\item{\code{GRID}}{Select whether to gather gridded or observed data
locations. See the 'AggInputs' class for more information on the 'GRID'
option. Provide either "grid" or "obs".}

\item{\code{dat_used}}{Option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection.}

\item{\code{orig_file}}{The original file name of the data to gather.}

\item{\code{agg_dat}}{Aggregated data saved as a result of the fetchAggDat() method.}

\item{\code{raw_dat}}{Raw data saved as a result of the fetchRawDat() method.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObsOP-new}{\code{ObsOP$new()}}
\item \href{#method-ObsOP-fetchAggDat}{\code{ObsOP$fetchAggDat()}}
\item \href{#method-ObsOP-fetchRawDat}{\code{ObsOP$fetchRawDat()}}
\item \href{#method-ObsOP-krigeDat}{\code{ObsOP$krigeDat()}}
\item \href{#method-ObsOP-calcNR}{\code{ObsOP$calcNR()}}
\item \href{#method-ObsOP-plotObsMaps}{\code{ObsOP$plotObsMaps()}}
\item \href{#method-ObsOP-plotScatters}{\code{ObsOP$plotScatters()}}
\item \href{#method-ObsOP-plotBoxplots}{\code{ObsOP$plotBoxplots()}}
\item \href{#method-ObsOP-plotViolins}{\code{ObsOP$plotViolins()}}
\item \href{#method-ObsOP-plotHistogram}{\code{ObsOP$plotHistogram()}}
\item \href{#method-ObsOP-clone}{\code{ObsOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-new}{}}}
\subsection{Method \code{new()}}{
Use this class to import data from an OFPE database, save figures and
explore observed data. This class can be initialized with a arguments required
for executing the methods of the class or with only a 'create' parameter,
see below. If the user supplies all arguments to the class, they can
typically use the methods with one or two arguments. A database connection
must be supplied to import data from the database. If the user imports data,
it is initialized into the class.

The logical argument, 'create', gives the user a choice to select whether to save
figures and maps. Depending on this choice, the output folders will be generated.
This is done by a private method that sets up the output location for
the figures that the model produces. This will not overwrite any previously
generated diagnostic or validation plots from the ModClass. Functions for
plots and maps all have save options which must be accompanied by a folder
path. The folder created is named 'Outputs'. This folder contains a
folder called 'ObsMaps' which contains maps of observed data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$new(
  out_path = NULL,
  farmername = NULL,
  fieldname = NULL,
  year = NULL,
  dat_tab = NULL,
  GRID = NULL,
  db = NULL,
  dat_used = NULL,
  utm_fieldname = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{out_path}}{Optional, path to folder to save outputs. Also serveds
as a logical for whether to create the 'Output' folder and save
figures, maps, and tables. If this argument is left null, the user will
have to provide the path to save outputs for each individual method.}

\item{\code{farmername}}{Optional, the name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{fieldname}}{Optional, the name of the field(s) to gather data for. Must
match the format in the OFPE database.}

\item{\code{year}}{Optional, the year to gather data for.}

\item{\code{dat_tab}}{Optional, the table that the data is in within the aggregated schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'pro', or 'sat').}

\item{\code{GRID}}{Optional, select whether to gather gridded or observed data
locations. See the 'AggInputs' class for more information on the 'GRID'
option. Provide either "grid" or "obs".}

\item{\code{db}}{Optional, connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{dat_used}}{Optional, option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection. Provide either
"decision_point" or "full_year".}

\item{\code{utm_fieldname}}{The name of the field (corresponding to the name in the
OFPE database) that should be used for identifying the UTM zone. In the
case of multiple fields (i.e. "sec1east & sec1west") choose one because
their UTM code will be the same.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Initialized R6 class and a folder created in the path for
model output figures if specified.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-fetchAggDat"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-fetchAggDat}{}}}
\subsection{Method \code{fetchAggDat()}}{
Gather data from an aggregated table in an
OFPE database. Provide the farmername, fieldname(s), year(s),
and type of data to gather. Requires database connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$fetchAggDat(
  db = self$db,
  farmername = self$farmername,
  fieldname = self$fieldname,
  year = self$year,
  dat_tab = self$dat_tab,
  GRID = self$GRID,
  dat_used = self$dat_used,
  store = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{fieldname}}{The name of the field(s) to gather data for. Must
match the format in the OFPE database.}

\item{\code{year}}{The year(s) to gather data for.}

\item{\code{dat_tab}}{The table that the data is in within the aggregated schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'pro', or 'sat').}

\item{\code{GRID}}{Select whether to gather gridded or observed data
locations. See the 'AggInputs' class for more information on the 'GRID'
option. Provide either "grid" or "obs".}

\item{\code{dat_used}}{Optional, option for the length of year to use data in the analysis,
simulation, and prescription building steps. See the 'AggInputs' class
documentation for more information on the 'dat_used' selection. Provide either
"decision_point" or "full_year".}

\item{\code{store}}{Logical, whether to save the data internally in the class.
Default is TRUE. Access via rxClass$agg_dat. If FALSE will return dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with specified data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-fetchRawDat"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-fetchRawDat}{}}}
\subsection{Method \code{fetchRawDat()}}{
Gather data from a raw table in an
OFPE database. Provide the farmername, the type of data, and the
name of the original file for the data ('orig_file'). Requires database
connection.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$fetchRawDat(
  db = self$db,
  farmername = self$farmername,
  dat_tab = self$dat_tab,
  orig_file,
  store = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db}}{Connection to a database to gather data from. Must be
set up and filled in the OFPE specific format.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{dat_tab}}{The table that the data is in within the raw schema of
the database. This corresponds to the type of data gathered (i.e. 'yld',
'aa_n_poly', 'aa_sr', etc.).}

\item{\code{orig_file}}{The original file name of the data to gather.}

\item{\code{store}}{Logical, whether to save the data internally in the class.
Default is TRUE. Access via rxClass$raw_dat. If FALSE will return dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with specified data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-krigeDat"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-krigeDat}{}}}
\subsection{Method \code{krigeDat()}}{
Interpolate data using kriging. Takes two datasets and the
specified column name in the source data for the variable to be kriged
to the the target data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$krigeDat(source_dat, target_dat, var)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_dat}}{The data to interpolate from.}

\item{\code{target_dat}}{The data to krige data to.}

\item{\code{var}}{The variable to interpolate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data table with interpolated data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-calcNR"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-calcNR}{}}}
\subsection{Method \code{calcNR()}}{
Method for calculating net-return in a dataset. This method
looks for columns specified as 'yld' or 'pro' as well as 'aa_X' specifying
the experimental input variable. The user also specifies the economic
conditions used to calculate net-return. If there is no protein data in
the dataset, the protein premium/dockage specifications are not required.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$calcNR(
  dat,
  yld_col_name = NULL,
  pro_col_name = NULL,
  exp_col_name,
  CEXP,
  FC,
  ssAC,
  Bp,
  B0pd = NULL,
  B1pd = NULL,
  B2pd = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{The data frame or data table.}

\item{\code{yld_col_name}}{Character, name of column containing yield data. Leave
NULL to calculate net-return based on just protein. If NULL, pro_col_name
cannot also be NULL.}

\item{\code{pro_col_name}}{Character, name of column containing protein data. Leave
NULL to calculate net-return based on just yield. If NULL, yld_col_name
cannot also be NULL.}

\item{\code{exp_col_name}}{Character, name of the column for the experimental input
data. Required.}

\item{\code{CEXP}}{The cost of the experimental input.}

\item{\code{FC}}{Fixed costs ($/acre) associated with production, not including
the input of interest. This includes things like the cost of labor, fuel, etc.}

\item{\code{ssAC}}{The cost ($/acre) of using site-specific technology or variable rate
applications. For farmers that have variable rate technology this cost may be zero,
otherwise is the cost per acre to hire the equipment/operators with variable rate
technology.}

\item{\code{Bp}}{The base price corresponding to the price for the system
type selected by the user (i.e. conventional or organic).}

\item{\code{B0pd}}{The intercept for the protein premium/dockage equation.}

\item{\code{B1pd}}{The coefficient for protein in the protein premium/dockage
equation.}

\item{\code{B2pd}}{The coefficient for protein squared for the protein
premium/dockage equation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with new 'NR' column.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-plotObsMaps"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-plotObsMaps}{}}}
\subsection{Method \code{plotObsMaps()}}{
This method is for plotting maps of observed variables. The user
must supply the data and variable to map. Other arguments relate to
labeling the map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotObsMaps(
  dat,
  var,
  var_col_name,
  var_label,
  var_main_label = NULL,
  fieldname = self$unique_fieldname,
  year = self$year,
  SAVE = self$SAVE,
  farmername = self$farmername,
  out_path = self$out_path,
  db = self$dbCon$db,
  utm_fieldname = self$utm_fieldname,
  utm_zone = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with variables to map. Must include an x and y
column.}

\item{\code{var}}{The label of the variable to map. Used in figure name.}

\item{\code{var_col_name}}{The name of the column of the variable in the
supplied data ('dat').}

\item{\code{var_label}}{The label to be applied to the legend of the map
corresponding to the variable mapped.}

\item{\code{var_main_label}}{The main label to apply to the map.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{year}}{Year of the observed data.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{farmername}}{The name of the farmer that manages the field.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{db}}{Connection to the OFPE database to identify UTM zone. Optional, will try
and calculate without. Defaults to whatever was passed in to initialize ObsOP class.
Optional, but require 'utm_zone' if omitted.}

\item{\code{utm_fieldname}}{Name of the field for identifying the UTM zone. Defaults to whatever
was passed in to initialize ObsOP class. Optional, but require 'utm_zone' if omitted.}

\item{\code{utm_zone}}{The EPSG code for the UTM zone that contains the data. Optional if a
database connection is provided to look it up.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggmap' object and maps saved in the specified output folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-plotScatters"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-plotScatters}{}}}
\subsection{Method \code{plotScatters()}}{
This method is for creating a scatterplot of variables specified by the
user. The user specifies the x and y column to plot and a variable or
variables to color points by.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotScatters(
  dat,
  x_var,
  y_var,
  x_lab = NULL,
  y_lab = NULL,
  color_var = NULL,
  color_lab = NULL,
  main_label,
  out_path = self$out_path,
  save_label = NULL,
  SAVE = self$SAVE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with variables to plot. Must include the columns for
specified data.}

\item{\code{x_var}}{The column name of the variable to plot on the x axis.}

\item{\code{y_var}}{The column name of the variable to plot on the y axis.}

\item{\code{x_lab}}{The label to be applied to the x axis of the plot.}

\item{\code{y_lab}}{The label to be applied to the y axis of the plot.}

\item{\code{color_var}}{The variable or variables, passed in as a character string,
to color the data by. If left NULL no coloring is applied.}

\item{\code{color_lab}}{The label to be applied to the legend for the fill color.}

\item{\code{main_label}}{Title for the figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{save_label}}{The label to apply to the filename. The y, x, and color
variables will also be added to the filename. Optional, but if SAVE = TRUE
and save_lable = NULL 'main_label' will be used.}

\item{\code{SAVE}}{Logical, whether to save figure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scatterplot and saved in the specified output folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-plotBoxplots"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-plotBoxplots}{}}}
\subsection{Method \code{plotBoxplots()}}{
This method is for creating a boxplot of variables specified by the
user. The user specifies the x and y column to plot and a variable or
variables to color points by.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotBoxplots(
  dat,
  x_var,
  y_var,
  x_lab = NULL,
  y_lab = NULL,
  color_var = NULL,
  color_lab = NULL,
  main_label,
  out_path = self$out_path,
  save_label = NULL,
  SAVE = self$SAVE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with variables to plot. Must include the columns for
specified data.}

\item{\code{x_var}}{The column name of the variable to plot on the x axis.}

\item{\code{y_var}}{The column name of the variable to plot on the y axis.}

\item{\code{x_lab}}{The label to be applied to the x axis of the plot.}

\item{\code{y_lab}}{The label to be applied to the y axis of the plot.}

\item{\code{color_var}}{The variable or variables, passed in as a character string,
to color the data by. If left NULL no coloring is applied.}

\item{\code{color_lab}}{The label to be applied to the legend for the fill color.}

\item{\code{main_label}}{Title for the figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{save_label}}{The label to apply to the filename. The y, x, and color
variables will also be added to the filename. Optional, but if SAVE = TRUE
and save_lable = NULL 'main_label' will be used.}

\item{\code{SAVE}}{Logical, whether to save figure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A boxplot and saved in the specified output folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-plotViolins"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-plotViolins}{}}}
\subsection{Method \code{plotViolins()}}{
This method is for creating a violin plot of variables specified by the
user. The user specifies the x and y column to plot and a variable or
variables to color points by. The violin plot shows the probability
density of observations, and also contains a boxplot within the violin
to show the IQR.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotViolins(
  dat,
  x_var,
  y_var,
  x_lab = NULL,
  y_lab = NULL,
  color_var = NULL,
  color_lab = NULL,
  main_label,
  out_path = self$out_path,
  save_label = NULL,
  SAVE = self$SAVE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with variables to plot. Must include the columns for
specified data.}

\item{\code{x_var}}{The column name of the variable to plot on the x axis.}

\item{\code{y_var}}{The column name of the variable to plot on the y axis.}

\item{\code{x_lab}}{The label to be applied to the x axis of the plot.}

\item{\code{y_lab}}{The label to be applied to the y axis of the plot.}

\item{\code{color_var}}{The variable or variables, passed in as a character string,
to color the data by. If left NULL no coloring is applied.}

\item{\code{color_lab}}{The label to be applied to the legend for the fill color.}

\item{\code{main_label}}{Title for the figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{save_label}}{The label to apply to the filename. The y, x, and color
variables will also be added to the filename. Optional, but if SAVE = TRUE
and save_lable = NULL 'main_label' will be used.}

\item{\code{SAVE}}{Logical, whether to save figure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A violin plot and saved in the specified output folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-plotHistogram"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-plotHistogram}{}}}
\subsection{Method \code{plotHistogram()}}{
This method is for creating a histogram of a variable specified by the
user. The user specifies the column in the data holding the data of interest
and provides a lavel.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$plotHistogram(
  dat,
  x_var,
  x_lab = NULL,
  color_var = NULL,
  color_lab = NULL,
  main_label,
  out_path = self$out_path,
  save_label = NULL,
  SAVE = self$SAVE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with variables to plot. Must include the columns for
specified data.}

\item{\code{x_var}}{The column name of the variable to plot on the x axis.}

\item{\code{x_lab}}{The label to be applied to the x axis of the plot.}

\item{\code{color_var}}{The variable or variables, passed in as a character string,
to color the data by. If left NULL no coloring is applied.}

\item{\code{color_lab}}{The label to be applied to the legend for the fill color.}

\item{\code{main_label}}{Title for the figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{save_label}}{The label to apply to the filename. The y, x, and color
variables will also be added to the filename. Optional, but if SAVE = TRUE
and save_lable = NULL 'main_label' will be used.}

\item{\code{SAVE}}{Logical, whether to save figure.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scatterplot and saved in 'Outputs/Maps' folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObsOP-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObsOP-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObsOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
