% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OFPE.R
\docType{package}
\name{OFPE}
\alias{OFPE}
\title{OFPE: On-Farm Precision Experiments data management and analysis tools}
\description{
The OFPE package contains tools for supporting the OFPE data framework
developed at Montana State University. This package contains S3 and R6
objects for executing data management and analysis tasks, including;
creating an OFPE formatted spatial database, importing on-farm and
remotely sensed data, aggregating data from disparate sources, data
analysis and optimization of inputs, simulationg of management outcomes,
and generation of prescriptive application maps for optimized inputs.
}
\section{General Functions}{

DBCon
}

\section{Database Creation & Managements}{

See Vignette 1 for database management and updating examples.

BuildDB, ManageDB, ManageFarmers, ManageFarms, ManageFields
}

\section{Data Import}{

See the Vignettes 2A and 2B for examples importing data from farms and
Google Earth Engine.

The OFPE data import functions and R6 class generators...
ImportOF,

The GEE data import functions and R6 class generators...
}

\section{Data Aggregation}{

See Vignette 3 for examples aggregating data from disparate sources.

The OFPE data aggregation functions and R6 class generators...
}

\section{Data Analysis and Simulation}{

See Vignette 4 for an example on analyzing collected data and simulating
management outcomes.

The OFPE data analysis and simulation functions and R6 class generators...
}

\section{Prescription Generation}{

See Vignette 5 for an example on creating an optimized site-specific variable
rate input prescription.

The OFPE prescription creation functions and R6 class generators...
}

\seealso{
\url{https://sites.google.com/site/ofpeframework/}

\url{https://paulhegedus.github.io/OFPE-Website/}

\url{https://github.com/paulhegedus/OFPEDATA.git}
}
