% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeTempFuns.R
\name{removeTempFarmerTables}
\alias{removeTempFarmerTables}
\title{Remove temporary tables from a farmer's schemas}
\usage{
removeTempFarmerTables(db, farmername)
}
\arguments{
\item{db}{Connection to an OFPE formatted database,}

\item{farmername}{Name of the farmer corresponding to the schemas to search
in for removal of temporary tables.}
}
\value{
NULL, tables removed in database.
}
\description{
Removes any temporary tables from the schemas
in an OFPE formatted database of the specified farmer. .
}
