% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimOP.R
\name{SimOP}
\alias{SimOP}
\title{R6 Class for plotting outputs of the OFPE simulation.}
\description{
R6 class for plotting outputs of the OFPE simulation such as
figures comparing management scenarios, probability tables, and maps.

This class should be initialized and executed after the user has initialized
and executed all other classes and has ran the SimClass$executeSim() method.
These figures rely on outputs from the simulation, so will cause errors if
used prior to execution of the simulation.

See the SimClass$executeSim() method documentation for more detail on the management
scenarios reported. NR is net-return. The abbreviations used are; NRmin: uniform minimum
rate ('AAmin'), NRfs: uniform farmer selected rate, NRact: if experiment applied in
simulated year, NRssopt: site-specific optimum rates, NRffopt: uniform full-
field optimum rate, NRopp: if the crop received the opposite system crop price (e.g.
'conventional' vs. 'organic' or reverse, see SimClass$executeSim() method
documentation for specifics on the difference in the calculation of this net-return
between systems).

Error checking inputs to these functions is scant because the user should only use
these after executing the simulation method in the 'SimClass' R6 class to generate
the necessary data for the output generation methods.
}
\seealso{
\code{\link{SimClass}} for the class that executes the Monte Carlo
simulation and provides the data for saving and plotting in this class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simClass}}{'SimClass' R6 object that has been initialized and the OFPE
simulation has been executed. This class holds the information necessary
to produce the output figures, maps, and tables from the Monte Carlo
simulation.}

\item{\code{input_list}}{Optional, list of inputs required for generating the
output figures. Recreation of relevant info from a 'SimClass' object. Only
needed if a 'SimClass' object is not passed into class.

'input_list' is a named list with the following fields;

dat_path: The path to where the data from the simulation was exported.
Must contain 'Bp.var', 'NRffmax', and 'NRopt' tables exported from the
'SimClass'. This class requires file names to be in the same format as
exported from the 'SimClass' to import.

unique_fieldname: Unique name for the field(s) used in the simulation.
This is used for saving and labeling of outputs. e.g. "sec35middle" or
"sec1east & sec1west". This must match the 'unique_fieldname' used to
save the simulation outputs.

unique_fxn: The function name generated by the 'SimClass' object that
was used for saving the simulation outputs. e.g. "yldGAM-proGAM".

sim_years: Years that the simulation was run. The output tables from
the simulation must be labeled with these years as exported from the
'SimClass' object.

opt: The optimization method used in the simulation. Used to identify
and label data. Must be in the same format as from the 'SimClass', e.g.
"deriv".

fieldsize: Size of the field in acres.

fs: The uniform rate of the experimental input that the farmer would
have applied in the absence of variable rate application.

expvar: The OFPE experimental input code for the experimental variable,
i.e. "aa_n" for as-applied nitrogen fertilizer or "aa_sr" for as-applied
seed rate.

farmername: The name of the farmer, corresponding to the format in the
database, e.g. lowercase.

respvar: Character vector of OFPE response variable codes that were
used in the model fitting and simulation processes, e.g. "yld" or "pro"
or 'c("yld", "pro")'.

db: Connection to an OFPE formatted database.

utm_fieldname: The name of the field (corresponding to the name in the
OFPE database) that should be used for identifying the UTM zone. In the
case of multiple fields (i.e. "sec1east & sec1west") choose one because
their UTM code will be the same.

out_path: Path to the folder to create an 'Outputs' folder and save
figures and tables.}

\item{\code{out_path}}{Output path to create folder for outputs. Taken from either
a 'SimClass' object or an optional 'input_list' field upon initialization.}

\item{\code{SAVE}}{Whether to save simulation outputs If NULL uses the user selected
choice. If not NULL and is logical, argument replaces previously set SAVE
options for the entire class.}

\item{\code{dat_path}}{The path to where the data from the simulation was exported.
Must contain 'Bp.var', 'NRffmax', and 'NRopt' tables exported from the
'SimClass'. This class requires file names to be in the same format as
exported from the 'SimClass' to import.}

\item{\code{unique_fieldname}}{Unique name for the field(s) used in the simulation.
This is used for saving and labeling of outputs. e.g. "sec35middle" or
"sec1east & sec1west". This must match the 'unique_fieldname' used to
save the simulation outputs.}

\item{\code{unique_fxn}}{The function name generated by the 'SimClass' object that
was used for saving the simulation outputs. e.g. "yldGAM-proGAM".}

\item{\code{sim_years}}{Years that the simulation was run. The output tables from
the simulation must be labeled with these years as exported from the
'SimClass' object.}

\item{\code{opt}}{The optimization method used in the simulation. Used to identify
and label data. Must be in the same format as from the 'SimClass', e.g. "deriv".}

\item{\code{fieldsize}}{Size of the field in acres.}

\item{\code{fs}}{The uniform rate of the experimental input that the farmer would
have applied in the absence of variable rate application.}

\item{\code{expvar}}{The OFPE experimental input code for the experimental variable,
i.e. "aa_n" for as-applied nitrogen fertilizer or "aa_sr" for as-applied
seed rate.}

\item{\code{farmername}}{The name of the farmer, corresponding to the format in the
database, e.g. lowercase.}

\item{\code{respvar}}{Character vector of OFPE response variable codes that were
used in the model fitting and simulation processes, e.g. "yld" or "pro"
or 'c("yld", "pro")'.}

\item{\code{db}}{Connection to an OFPE formatted database.}

\item{\code{utm_fieldname}}{The name of the field (corresponding to the name in the
OFPE database) that should be used for identifying the UTM zone. In the
case of multiple fields (i.e. "sec1east & sec1west") choose one because
their UTM code will be the same.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimOP$new()}}
\item \href{#method-savePlots}{\code{SimOP$savePlots()}}
\item \href{#method-plotNRbox}{\code{SimOP$plotNRbox()}}
\item \href{#method-mgmtNRprobTable}{\code{SimOP$mgmtNRprobTable()}}
\item \href{#method-plotNRbar}{\code{SimOP$plotNRbar()}}
\item \href{#method-plotTotExpAppl}{\code{SimOP$plotTotExpAppl()}}
\item \href{#method-plotSSOPThist}{\code{SimOP$plotSSOPThist()}}
\item \href{#method-plotFFOPThist}{\code{SimOP$plotFFOPThist()}}
\item \href{#method-plotSimMaps}{\code{SimOP$plotSimMaps()}}
\item \href{#method-loadBpVar}{\code{SimOP$loadBpVar()}}
\item \href{#method-loadNRffmax}{\code{SimOP$loadNRffmax()}}
\item \href{#method-makeTF4}{\code{SimOP$makeTF4()}}
\item \href{#method-loadNRopt}{\code{SimOP$loadNRopt()}}
\item \href{#method-clone}{\code{SimOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Use this class to save figures and tables from simulations performed using
the 'SimClass' R6 object. This class can be initialized with an executed
'SimClass' object, or a list of inputs. See 'input_list' in this class'
description. This class can also be initialized with neither and methods
for generating individual figures and tables can be executed by passing
in arguments. When used in this manner this may be more susceptible to
errors as the user is required to specify arguments in the same format
as found in the 'SimClass' object.

Also pass in a logical to the 'create' argument to select whether to save
figures. Depending on this choice, the output folder will be generated.
This is done by a private method that sets up the output location for
the figures that the model produces. This will not overwrite any previously
generated diagnostic or validation plots from the ModClass. Functions for
plots and maps all have save options which must be accompanied by a folder
path. The folder created is named 'Outputs'. This folder contains a
folder called 'Maps' which contains maps of estimated responses, net-return,
and site-specific optimized rates. Another folder called 'EXP' is created
that holds figures related to the amounts and distributions for as-applied
rates of the experimental variable. Within this folder, subfolders called
'ffoptEXP', 'ssoptEXP', and 'EXPapplied' are created. The acronyms SSOPT and
FFOPT correspond to site-specific optimized and full-field optimized rates.
Another main folder named 'NR' is created to hold net-return figures comparing
management outcomes. Subfolders within 'NR' are 'NRboxplots', 'NRbarplots'
and 'NRprobabilities'. The two former contain figures, and the latter contains
a table with the probability that the SSOPT strategy yields a higher net-return
than each of the other management strategies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$new(simClass = NULL, input_list = NULL, create = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simClass}}{Optional, 'SimClass' R6 object that has been initialized and the OFPE
simulation has been executed. This class holds the information necessary
to produce the output figures, maps, and tables from the Monte Carlo
simulation.}

\item{\code{input_list}}{Optional, list of inputs required for generating the
output figures. Recreation of relevant info from a 'SimClass' object. Only
needed if a 'SimClass' object is not passed into class.}

\item{\code{create}}{create Logical, whether to create the 'Output' folder and save
figures, maps, and tables. Default is TRUE, pass FALSE to skip output folder
step and prevent any output generation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A folder created in the path for model output figures.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savePlots"></a>}}
\if{latex}{\out{\hypertarget{method-savePlots}{}}}
\subsection{Method \code{savePlots()}}{
Method for saving simulation output plots (see below). This method is used
to call the public methods for each figure, map,  and table generated. This
applies over the user selected simulation years. The user can pass a
logical argument to this function which will disable all plotting and
not save anything to your computer It will override any previous
setting you have set for 'SAVE'. If calling plots individually, a
SAVE option for each can be selected.

This method will only work if the 'SimOP' class has been initialized with
a 'SimClass' or 'input_list' object. If the 'SimOP' class is initialized with
neither, than individual plotting functions must be executed with user
supplied inputs. This class uses internal fields in the 'SimOP' class that
are created on instantiation with a 'SimClass' or 'input_list' object.
class from these objects.

The output methods and output created by this method are; 'plotNRbox',
'plotNRbar', 'mgmtNRprobTable', 'plotFFOPThist', 'plotTotExpAppl',
and 'plotSSOPThist'. Maps are created with the 'plotSimMaps' methods.
The maps saved are net-return maps for the SSOPT, FFOPT, FS, and Act
management strategies, estimated yield and protein maps for the SSOPT
and FS management strategies, and the SSOPT experimental rate map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$savePlots(SAVE = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SAVE}}{Whether to save simulation outputs If NULL uses the user selected
choice. If not NULL and is logical, argument replaces previously set SAVE
options for the entire class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Simulation outputs in the 'Outputs' folder.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNRbox"></a>}}
\if{latex}{\out{\hypertarget{method-plotNRbox}{}}}
\subsection{Method \code{plotNRbox()}}{
Convert the Bp.var matrix to a dataframe for plotting. Reorder the levels
of the method for plotting. Now, start to save the tables and plots showing
results from the simulations. First, the variation in net returns over
different price years.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotNRbox(
  Bp.var,
  fieldname = self$fieldname,
  fxn = self$unique_fxn,
  sim_year,
  opt = self$opt,
  SAVE = self$SAVE,
  out_path = self$out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggplot' object and data saved in 'Outputs/NR/NRboxplots' if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mgmtNRprobTable"></a>}}
\if{latex}{\out{\hypertarget{method-mgmtNRprobTable}{}}}
\subsection{Method \code{mgmtNRprobTable()}}{
Calculate the propabability that a given management method is more profitable
than the other management strategies. For example, compare the probability that
the site-specific optimization strategy yielded a higher net-retrun than the
other management strategies. This probability is generated from across all
iterations of the Monte Carlo simulation, so reflects the variation in economic
conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$mgmtNRprobTable(
  mgmt,
  Bp.var,
  rt,
  fieldname = self$unique_fieldname,
  fxn = self$unique_fxn,
  sim_year,
  opt = self$opt,
  SAVE = self$SAVE,
  out_path = self$out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mgmt}}{Select the management strategy to compare, select from; 'NR.ssopt',
'NR.min', 'NR.fs', 'NR.ffopt', 'NR.act', 'NR.opp'. See 'SimClass' documentation
or this class' description for acronym definitions.}

\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{rt}}{Number of times to sample the data with replacement to generate the
probability that the net-return of the selected strategy was higher than each
other strategy.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame and data saved in 'Outputs/NR/NRProbabilities' if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNRbar"></a>}}
\if{latex}{\out{\hypertarget{method-plotNRbar}{}}}
\subsection{Method \code{plotNRbar()}}{
Plot the mean net return for each management strategy as a
bar plot and a quarter standard deviation around each mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotNRbar(
  Bp.var,
  fieldname = self$unique_fieldname,
  fxn = self$unique_fxn,
  sim_year,
  opt = self$opt,
  SAVE = self$SAVE,
  out_path = self$out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggplot' object and data saved in 'Outputs/NR/NRbarplots' if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotTotExpAppl"></a>}}
\if{latex}{\out{\hypertarget{method-plotTotExpAppl}{}}}
\subsection{Method \code{plotTotExpAppl()}}{
Create a bar plot of the total experimental input applied with
each strategy. The amount of the site-specific input applied
is averaged across the simulation for each point and summed to
get a total amount of experimental input applied with the site-
specific method. The total amount of experimental input applied
with all other strategies is calculated the same way.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotTotExpAppl(
  TF4,
  expvar = self$expvar,
  fieldname = self$unique_fieldname,
  fxn = self$unique_fxn,
  sim_year,
  opt = self$opt,
  SAVE = self$SAVE,
  out_path = self$out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TF4}}{Data.frame with columns for the amount of experimental rates applied
for each strategy.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggplot' object and data saved in 'Outputs/EXP/EXPapplied' if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSSOPThist"></a>}}
\if{latex}{\out{\hypertarget{method-plotSSOPThist}{}}}
\subsection{Method \code{plotSSOPThist()}}{
Generate a histogram of site specific experimental inptut rates within
a field. Uses the average site-specific rate for each point across the
simulation, indicating the mean optimum rate at each point across all
economic conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotSSOPThist(
  NRopt,
  TF4,
  expvar = self$expvar,
  fieldname = self$unique_fieldname,
  fxn = self$unique_fxn,
  sim_year,
  opt = self$opt,
  SAVE = self$SAVE,
  out_path = self$out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NRopt}}{Data frame with the net-returns and experimental optimums
for every point for every simulation iteration.}

\item{\code{TF4}}{Data.frame with columns for the amount of experimental rates applied
for each strategy.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggplot' object and data saved in 'Outputs/EXP/ssoptEXP' if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFFOPThist"></a>}}
\if{latex}{\out{\hypertarget{method-plotFFOPThist}{}}}
\subsection{Method \code{plotFFOPThist()}}{
Generate a histogram of full-field experimental input rates identified
in the simulation. Shows the distribution of full-field optimum rates
across all economic scenarios in the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotFFOPThist(
  Bp.var,
  TF4,
  expvar = self$expvar,
  fieldname = self$unique_fieldname,
  fxn = self$unique_fxn,
  sim_year,
  opt = self$opt,
  SAVE = self$SAVE,
  out_path = self$out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{TF4}}{Data.frame with columns for the amount of experimental rates applied
for each strategy.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggplot' object and data saved in 'Outputs/EXP/ffoptEXP' if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSimMaps"></a>}}
\if{latex}{\out{\hypertarget{method-plotSimMaps}{}}}
\subsection{Method \code{plotSimMaps()}}{
This method is for plotting maps of simulation outcomes. These
include the net-returns from the management strategies, predicted
responses, and optimized rates. Is a wrapper to OFPE::plotMaps().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotSimMaps(
  dat,
  var,
  var_col_name,
  var_label,
  var_main_label,
  fieldname = self$unique_fieldname,
  sim_year,
  fxn = self$unique_fxn,
  opt = self$opt,
  SAVE = self$SAVE,
  farmername = self$farmername,
  out_path = self$out_path,
  db = self$db,
  utm_fieldname = self$utm_fieldname
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with the net-returns and experimental optimums
for every point for every simulation iteration.}

\item{\code{var}}{The label of the variable to map. Used in figure name.}

\item{\code{var_col_name}}{The name of the column of the variable in the
supplied data ('dat').}

\item{\code{var_label}}{The label to be applied to the legend of the map
corresponding to the variable mapped.}

\item{\code{var_main_label}}{The main label to apply to the map.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{opt}}{The optimization method used in the simulation. For labeling.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{farmername}}{The name of the farmer that manages the field.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}

\item{\code{db}}{Connection to the OFPE database to identify UTM zone.}

\item{\code{utm_fieldname}}{Name of the field for identifying the UTM zone.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 'ggmap' object and maps saved in 'Outputs/Maps' folder if selected.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadBpVar"></a>}}
\if{latex}{\out{\hypertarget{method-loadBpVar}{}}}
\subsection{Method \code{loadBpVar()}}{
Gets the Bp.var output table from the 'SimClass' simulation. The Bp.var table contains means
across each field for each iteration of the simulation. This records base
price received, the cost of the input, the mean net-return per acre for
each management strategy, and the full-field optimum rate. Arguments are
used to fill in the file name in the same format as how the file was
saved from the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$loadBpVar(
  dat_path = self$dat_path,
  unique_fieldname = self$unique_fieldname,
  unique_fxn = self$unique_fxn,
  sim_year,
  opt = self$unique_fxn
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat_path}}{The path to the folder where the data is saved.}

\item{\code{unique_fieldname}}{Unique field name corresponding to the fieldname
in the file name.}

\item{\code{unique_fxn}}{The functional form of the models used for analysis
found in the file name.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used. Found in the file name.}

\item{\code{opt}}{The optimization method used in the simulation. Found in file
name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with 10 columns and rows equal to the number of
iterations in the simulation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadNRffmax"></a>}}
\if{latex}{\out{\hypertarget{method-loadNRffmax}{}}}
\subsection{Method \code{loadNRffmax()}}{
Gets the NRffmax table that is an output from the 'SimClass' simulation with rows for
each iteration and the full-field optimum experimental input rate and the
total net-return from the full-field. Arguments are used to fill in the file
name in the same format as how the file was saved from the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$loadNRffmax(
  dat_path = self$dat_path,
  unique_fieldname = self$unique_fieldname,
  unique_fxn = self$unique_fxn,
  sim_year,
  opt = self$opt
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat_path}}{The path to the folder where the data is saved.}

\item{\code{unique_fieldname}}{Unique field name corresponding to the fieldname
in the file name.}

\item{\code{unique_fxn}}{The functional form of the models used for analysis
found in the file name.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used. Found in the file name.}

\item{\code{opt}}{The optimization method used in the simulation. Found in file
name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with 3 columns and rows equal to the number of
iterations in the simulation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-makeTF4"></a>}}
\if{latex}{\out{\hypertarget{method-makeTF4}{}}}
\subsection{Method \code{makeTF4()}}{
Create the table TF4, which is used for the output functions 'plotFFOPThist',
'plotTotExpAppl', and 'plotSSOPThist'. This contains mean values for each
of the strategies to create bar charts of experimental inputs used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$makeTF4(NRopt, NRffmax, fieldsize = self$fieldsize, fs = self$fs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NRopt}}{'SimClass' output table that contains every point with
the net-return and experimenal rates for each management strategy, and
the estimated crop responses for each location for the simulation year that
are the means across all simulations.}

\item{\code{NRffmax}}{'SimClass' output table that contains the optimum full-field
experimental rate, the total net-return for each iteration of the simulation.}

\item{\code{fieldsize}}{Size of the field in acres.}

\item{\code{fs}}{The farmer selected experimental rate used in the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with SSOPT and FFOPT experimental input results for
each management strategy.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loadNRopt"></a>}}
\if{latex}{\out{\hypertarget{method-loadNRopt}{}}}
\subsection{Method \code{loadNRopt()}}{
Gets the NRopt table that is an output from the 'SimClass' simulation with rows for
each in the field. Columns contain the means across all iterations of the
simulation. Columns in this file are the  price received, cost of nitrogen,
SSOPT and FFOPT experimental rates and the estimated net-return, yield, and
protein for each management strategy at each location in the field.
Arguments are used to fill in the file name in the same format as how the file was saved from the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$loadNRopt(
  dat_path = self$dat_path,
  unique_fieldname = self$unique_fieldname,
  unique_fxn = self$unique_fxn,
  sim_year,
  opt = self$opt
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat_path}}{The path to the folder where the data is saved.}

\item{\code{unique_fieldname}}{Unique field name corresponding to the fieldname
in the file name.}

\item{\code{unique_fxn}}{The functional form of the models used for analysis
found in the file name.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used. Found in the file name.}

\item{\code{opt}}{The optimization method used in the simulation. Found in file
name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data frame with 22 columns and rows equal to the number of
observations in the dataset.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
