% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimOP.R
\name{SimOP}
\alias{SimOP}
\title{R6 Class for plotting outputs of the OFPE simulation.}
\description{
R6 class for plotting outputs of the OFPE simulation such as
figures comparing management scenarios, probability tables, and maps.

This class should be initialized and executed after the user has initialized
and executed all other classes and has ran the SimClass$executeSim() method.
These figures rely on outputs from the simulation, so will cause errors if
used prior to execution of the simulation.

See the SimClass$executeSim() method documentation for more detail on the management
scenarios reported. NR is net-return. The abbreviations used are; NRmin: uniform minimum
rate ('AAmin'), NRfs: uniform farmer selected rate, NRact: if experiment applied in
simulated year, NRssopt: site-specific optimum rates, NRffopt: uniform full-
field optimum rate, NRopp: if the crop received the opposite system crop price (e.g.
'conventional' vs. 'organic' or reverse, see SimClass$executeSim() method
documentation for specifics on the difference in the calculation of this net-return
between systems).

Error checking inputs to these functions is scant because the user should only use
these after executing the simulation method in the SimClass R6 class to generate
the necessary data for the output generation methods.
}
\seealso{
\code{\link{SimClass}} for the class that executes the Monte Carlo
simulation and provides the data for saving and plotting in this class.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{simClass}}{SimClass R6 object that has been initialized and the OFPE
simulation has been executed. This class holds the information necessary
to produce the output figures, maps, and tables from the Monte Carlo
simulation.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimOP$new()}}
\item \href{#method-savePlots}{\code{SimOP$savePlots()}}
\item \href{#method-plotNRbox}{\code{SimOP$plotNRbox()}}
\item \href{#method-mgmtNRprobTable}{\code{SimOP$mgmtNRprobTable()}}
\item \href{#method-plotNRbar}{\code{SimOP$plotNRbar()}}
\item \href{#method-plotTotExpAppl}{\code{SimOP$plotTotExpAppl()}}
\item \href{#method-plotSSOPThist}{\code{SimOP$plotSSOPThist()}}
\item \href{#method-plotFFOPThist}{\code{SimOP$plotFFOPThist()}}
\item \href{#method-plotSimMaps}{\code{SimOP$plotSimMaps()}}
\item \href{#method-plotActNR}{\code{SimOP$plotActNR()}}
\item \href{#method-clone}{\code{SimOP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Pass in a SimClass R6 object that has results from an executed simulation.
Also pass in a logical to 'create' to override any previously selected
SAVE options form the SimClass object. Depending on this choice, the output
folder will be generated. This is done by a private method that sets up the
output location for the figures that the model produces. This will not
overwrite any previously generated diagnostic or validation plots from the
ModClass. The folder created is named 'Outputs'. This folder contains a
folder called 'Maps' which contains maps of estimated responses, net-return,
and site-specific optimized rates. Another folder called 'EXP' is created
that holds figures related to the amounts and distributions for as-applied
rates of the experimental variable. Within this folder, subfolders called
'ffoptEXP', 'ssoptEXP', and 'EXPapplied' are created. The acronyms SSOPT and
FFOPT correspond to site-specific optimized and full-field optimized rates.
Another main folder named 'NR' is created to hold net-return figures comparing
management outcomes. Subfolders within 'NR' are 'NRboxplots', 'NRbarplots'
and 'NRprobabilities'. The two former contain figures, and the latter contains
a table with the probability that the SSOPT strategy yields a higher net-return
than each of the other management strategies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$new(simClass, create = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{simClass}}{SimClass R6 object that has been initialized and the OFPE
simulation has been executed. This class holds the information necessary
to produce the output figures, maps, and tables from the Monte Carlo
simulation.}

\item{\code{create}}{create Logical, whether to create the 'Output' folder and save
figures, maps, and tables. Default is TRUE, pass FALSE to skip output folder
step and prevent any output generation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A folder created in the path for model output figures.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-savePlots"></a>}}
\if{latex}{\out{\hypertarget{method-savePlots}{}}}
\subsection{Method \code{savePlots()}}{
Method for saving simulation output plots. This method is used to call
the public methods for each figure, map,  and table generated. This
applies over the user selected simulation years. The user can pass a
logical argument to this function which will disable all plotting and
not save anything to your computer It will override any previous
setting you have set for 'SAVE'. If calling plots individually, a
SAVE option for each can be selected.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$savePlots(SAVE = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{SAVE}}{Whether to save simulation outputs If NULL uses the user selected
choice. If not NULL and is logical, argument replaces previously set SAVE
options for the entire class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Simulation outputs in the 'Outputs' folder.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNRbox"></a>}}
\if{latex}{\out{\hypertarget{method-plotNRbox}{}}}
\subsection{Method \code{plotNRbox()}}{
Convert the Bp.var matrix to a dataframe for plotting. Reorder the levels
of the method for plotting. Now, start to save the tables and plots showing
results from the simulations. First, the variation in net returns over
different price years.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotNRbox(Bp.var, fieldname, fxn, sim_year, opt, SAVE, out_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved in 'Outputs/NR/NRboxplots'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mgmtNRprobTable"></a>}}
\if{latex}{\out{\hypertarget{method-mgmtNRprobTable}{}}}
\subsection{Method \code{mgmtNRprobTable()}}{
Calculate the propabability that a given management method is more profitable
than the other management strategies. For example, compare the probability that
the site-specific optimization strategy yielded a higher net-retrun than the
other management strategies. This probability is generated from across all
iterations of the Monte Carlo simulation, so reflects the variation in economic
conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$mgmtNRprobTable(
  mgmt,
  Bp.var,
  rt,
  fieldname,
  fxn,
  sim_year,
  opt,
  SAVE,
  out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mgmt}}{Select the management strategy to compare, select from; 'NR.ssopt',
'NR.min', 'NR.fs', 'NR.ffopt', 'NR.act', 'NR.opp'. See SimClass documentation
or this class' description for acronym definitions.}

\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{rt}}{Number of times to sample the data with replacement to generate the
probability that the net-return of the selected strategy was higher than each
other strategy.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved in 'Outputs/NR/NRProbabilities'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotNRbar"></a>}}
\if{latex}{\out{\hypertarget{method-plotNRbar}{}}}
\subsection{Method \code{plotNRbar()}}{
Plot the mean net return for each management strategy as a
bar plot and a quarter standard deviation around each mean.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotNRbar(Bp.var, fieldname, fxn, sim_year, opt, SAVE, out_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved in 'Outputs/NR/NRbarplots'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotTotExpAppl"></a>}}
\if{latex}{\out{\hypertarget{method-plotTotExpAppl}{}}}
\subsection{Method \code{plotTotExpAppl()}}{
Create a bar plot of the total experimental input applied with
each strategy. The amount of the site-specific input applied
is averaged across the simulation for each point and summed to
get a total amount of experimental input applied with the site-
specific method. The total amount of experimental input applied
with all other strategies is calculated the same way.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotTotExpAppl(
  TF4,
  expvar,
  fieldname,
  fxn,
  sim_year,
  opt,
  SAVE,
  out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{TF4}}{Data.frame with columns for the amount of experimental rates applied
for each strategy.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved in 'Outputs/EXP/EXPapplied'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSSOPThist"></a>}}
\if{latex}{\out{\hypertarget{method-plotSSOPThist}{}}}
\subsection{Method \code{plotSSOPThist()}}{
Generate a histogram of site specific experimental inptut rates within
a field. Uses the average site-specific rate for each point across the
simulation, indicating the mean optimum rate at each point across all
economic conditions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotSSOPThist(
  NRopt,
  TF4,
  expvar,
  fieldname,
  fxn,
  sim_year,
  opt,
  SAVE,
  out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{NRopt}}{Data frame with the net-returns and experimental optimums
for every point for every simulation iteration.}

\item{\code{TF4}}{Data.frame with columns for the amount of experimental rates applied
for each strategy.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved in 'Outputs/EXP/ssoptEXP'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotFFOPThist"></a>}}
\if{latex}{\out{\hypertarget{method-plotFFOPThist}{}}}
\subsection{Method \code{plotFFOPThist()}}{
Generate a histogram of full-field experimental input rates identified
in the simulation. Shows the distribution of full-field optimum rates
across all economic scenarios in the simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotFFOPThist(
  Bp.var,
  TF4,
  expvar,
  fieldname,
  fxn,
  sim_year,
  opt,
  SAVE,
  out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Bp.var}}{Data.table containing the average net-returns from the field for each
management type for every iteration of the simulation.}

\item{\code{TF4}}{Data.frame with columns for the amount of experimental rates applied
for each strategy.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Data saved in 'Outputs/EXP/ffoptEXP'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotSimMaps"></a>}}
\if{latex}{\out{\hypertarget{method-plotSimMaps}{}}}
\subsection{Method \code{plotSimMaps()}}{
This method is for plotting maps of simulation outcomes. These
include the net-returns from the management strategies, predicted
responses, and optimized rates. Is a wrapper to OFPE::plotMaps().
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotSimMaps(
  dat,
  var,
  var_col_name,
  var_label,
  var_main_label,
  fieldname,
  sim_year,
  fxn,
  opt = NULL,
  SAVE = TRUE,
  farmername,
  out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data frame with the net-returns and experimental optimums
for every point for every simulation iteration.}

\item{\code{var}}{The label of the variable to map. Used in figure name.}

\item{\code{var_col_name}}{The name of the column of the variable in the
supplied data ('dat').}

\item{\code{var_label}}{The label to be applied to the legend of the map
corresponding to the variable mapped.}

\item{\code{var_main_label}}{The main label to apply to the map.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{sim_year}}{Year that the simulation was performed for. Indicates the
weather conditions used.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{opt}}{The optimization method used in the simulation.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{farmername}}{The name of the farmer that manages the field.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Maps saved in 'Outputs/Maps'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotActNR"></a>}}
\if{latex}{\out{\hypertarget{method-plotActNR}{}}}
\subsection{Method \code{plotActNR()}}{
Function for creating a map of the actual net-return observed in the
field for a selected year available in the data the user specified
for fitting the models (i.e. if you used multiple years to fit a model
you can map the net-return for either year).

This function uses the observed data for the specified year and if
there is multiple response variables, predicts them to the yield
dataset because this is more spatially dense. This uses the model
fit in the analysis step that was used in the simulation. After
both responses are in the same dataset, net-return is calculated
based on the actual as-applied experimental input and mapped across
the field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$plotActNR(
  dat,
  year,
  Bp,
  B0pd,
  B1pd,
  B2pd,
  CEXP,
  FC,
  ssAC,
  respvar,
  expvar,
  modClass,
  fieldname,
  farmername,
  fxn,
  SAVE,
  out_path
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{Data.table or data.frame to use for calculating net-return.
Must include all covariates used in the model fitting processes. These
are the locations actual net-return is calculated for.}

\item{\code{year}}{A single year, present in 'dat', for which to calculate and
map the actual net-return.}

\item{\code{Bp}}{The base price corresponding to the price for the system
type selected by the user (i.e. conventional or organic).}

\item{\code{B0pd}}{The intercept for the protein premium/dockage equation.}

\item{\code{B1pd}}{The coefficient for protein in the protein premium/dockage
equation.}

\item{\code{B2pd}}{The coefficient for protein squared for the protein
premium/dockage equation.}

\item{\code{CEXP}}{The cost of the experimental input.}

\item{\code{FC}}{Fixed costs ($/acre) associated with production, not including
the input of interest. This includes things like the cost of labor, fuel, etc.}

\item{\code{ssAC}}{The cost ($/acre) of using site-specific technology or variable rate
applications. For farmers that have variable rate technology this cost may be zero,
otherwise is the cost per acre to hire the equipment/operators with variable rate
technology.}

\item{\code{respvar}}{Response variable(s) used to optimize experimental inputs based
off of. The user can select 'Yield' and/or 'Protein' based on data
availability.}

\item{\code{expvar}}{Experimental variable optimized, select/input
'As-Applied Nitrogen' or 'As-Applied Seed Rate'. This is the type of
input that was experimentally varied across the field as part of the
on-farm experimentation.}

\item{\code{modClass}}{ModClass R6 object that has been initialized and with
fitted models. The fitted models in this object are used to predict
crop responses at each location in the simulation data for all
experimental rates in the user specified range.}

\item{\code{fieldname}}{Unique field name corresponding to all fields used in the simulation.}

\item{\code{farmername}}{The name of the farmer that manages the field.}

\item{\code{fxn}}{The functional form of the models used for analysis.}

\item{\code{SAVE}}{Logical, whether to save figure.}

\item{\code{out_path}}{The path to the folder in which to store and
save outputs from the simulation.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Map of the actual net-return from a given year in 'Outputs/Maps'.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimOP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
