% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxFuns.R
\name{trimGrid}
\alias{trimGrid}
\title{Trim grid to field buffer}
\usage{
trimGrid(rx_sdt, fieldname, db, farmername)
}
\arguments{
\item{rx_sdt}{A 'sf' object containing the experiment or prescription
grid generated by user inputs.}

\item{fieldname}{Provide the name of the field for the experiment or prescription.}

\item{db}{Connection to an OFPE formatted database.}

\item{farmername}{Name of the farmer that owns or manages the field(s) that
an experiment or prescription is going to be generated for. Must be same for all
fields.}
}
\value{
A 'sf' object with a clipped grid for experimentation or prescription generation.
}
\description{
This function takes the experiment or prescription grid
from the database and trims it to the boundary of a 10m buffer within
the field(s).
}
